% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2rdat.r
\name{add2rdat}
\alias{add2rdat}
\title{Add elements to rdat}
\usage{
add2rdat(
  rdat,
  fleet_sel_abb_key = data.frame(F = c("cHL.D", "rHB.D", "rGN.D"), sel = c("D", "D",
    "D")),
  F_names_exclude = c("F.Fmsy", "F.full", "F.sum", "F.F30.ratio", "F.F40.ratio")
)
}
\arguments{
\item{rdat}{BAM output rdat (list) object read in with dget()}

\item{fleet_sel_abb_key}{Data frame for specifying which selectivities to use with time series of fishing mortality (F) for individual fleets
when the abbreviations in the rdat do not exactly match with a selectivity matrix. Column F lists the abbreviations F time series. Column sel
lists the abbreviations associated with the selectivity you want to use with the F time series. This is used to compute fleet-specific F-age
matrices (year,age).}

\item{F_names_exclude}{Character vector of names of columns in t.series that you want to exclude. The function identifies any columns
that begin with "F." The default value of \code{F_names_exclude} is a character vector of column names found in some BAM rdat files that
that begin with "F." but are not fleet-specific F time series.
to a single fleet.}
}
\description{
Add elements to rdat
}
\examples{
\dontrun{
# Add to any of the current BAM models
rdat_AtMe <- add2rdat(rdat_AtlanticMenhaden)
rdat_BlSB <- add2rdat(rdat_BlackSeaBass)
rdat_BlTi <- add2rdat(rdat_BluelineTilefish)
rdat_Cobi <- add2rdat(rdat_Cobia)
rdat_GagG <- add2rdat(rdat_GagGrouper,fleet_sel_abb_key=data.frame("F"=c("cHL.D","rHB.D","rGN.D"),"sel"=c("D","D","D")))
rdat_GrTr <- add2rdat(rdat_GrayTriggerfish)
rdat_GrAm <- add2rdat(rdat_GreaterAmberjack)
rdat_ReGr <- add2rdat(rdat_RedGrouper)
rdat_RePo <- add2rdat(rdat_RedPorgy)
rdat_ReSn <- add2rdat(rdat_RedSnapper)
rdat_SnGr <- add2rdat(rdat_SnowyGrouper)
rdat_Tile <- add2rdat(rdat_Tilefish)
rdat_VeSn <- add2rdat(rdat_VermilionSnapper,fleet_sel_abb_key=data.frame("F"=c("rGN.D"),"sel"=c("rHB.D")))
}
}
\author{
Nikolai Klibansky
}
\keyword{MSEtool}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
