% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agelenprob.r
\name{agelenprob}
\alias{agelenprob}
\title{agelenprob}
\usage{
agelenprob(
  age,
  len,
  binmid_lenc,
  cv_lenc,
  plot = FALSE,
  label = TRUE,
  label_min = 0.01
)
}
\arguments{
\item{age}{vector of ages}

\item{len}{vector of lengths corresponding to age}

\item{binmid_lenc}{vector of lengths corresponding to midpoints of length composition bins. These do not need to match len.}

\item{cv_lenc}{cv associated with variation in length at age associated with lenc}

\item{plot}{logical. If TRUE, plot an image of the resulting probability matrix}

\item{label}{logical. If TRUE, add probabilities as text to image plot}

\item{label_min}{numeric. Set minimum probability to label in plot}
}
\description{
Build age-length probability matrix associated with a set of length comps
}
\examples{
\dontrun{
rdat <- rdat_RedSnapper
agelenprob(
  age=rdat$a.series$age,
  len=rdat$a.series$length,
  binmid_lenc=as.numeric(colnames(rdat$comp.mats$lcomp.cHL.ob)),
  cv_lenc=rdat$parm.cons$len.cv.val.L[8],
  plot=TRUE)
}
}
\author{
Kyle Shertzer and Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
