% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to_r.r
\name{bam_to_r}
\alias{bam_to_r}
\title{BAM inputs to R object conversion function}
\usage{
bam_to_r(dat.file, tpl.file, cxx.file, L.init.user = NULL)
}
\arguments{
\item{dat.file}{dat file path}

\item{tpl.file}{tpl file path}

\item{cxx.file}{cxx file path}

\item{L.init.user}{user supplied L.init object. If missing, the function builds one from values in the dat and tpl files.}
}
\description{
This script reads in BAM dat, tpl, and cxx files, and converts them to R objects (character vectors of each line of the code). It also builds an R list on the tpl init objects, from the tpl and dat files.
}
\examples{
bam_to_r()
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
