% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bamStockMisc}
\alias{bamStockMisc}
\title{A dataset containing miscellaneous information related to species assessed with the Beaufort Assessment Model}
\format{
An object of class \code{"data.frame"}
\describe{
\item{Name}{Common name in Pascal case (e.g. BlackSeaBass)}
\item{Species}{Genus and species (e.g. Centropristis striata)}
\item{maxage}{Maximum age of species according to most recent stock assessment}
\item{herm}{Hermaphroditism: "gonochoristic" or "protogynous"}
}
}
\source{
Values are from current taxonomic standings and recent SEDAR stock assessment reports
}
\usage{
bamStockMisc
}
\description{
A dataset containing miscellaneous information related to species assessed with the Beaufort Assessment Model
}
\keyword{datasets}
