% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic2lk.r
\name{bic2lk}
\alias{bic2lk}
\alias{bic_to_lk}
\title{Convert Bayesian Information Criterion (BIC) to likelihood (lk) value}
\usage{
bic2lk(bic, k, n)

bic_to_lk(bic, k, n)
}
\arguments{
\item{bic}{Bayesian Information Criterion (BIC) value}

\item{k}{number of parameters estimated}

\item{n}{number of observations}
}
\description{
Converts Bayesian Information Criterion (BIC) to likelihood (lk) value
}
\examples{
\dontrun{
bic2lk(bic=10,k=3,n=100)
bic2lk(bic=10,k=5,n=100)
}

}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
