% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_to_lk.r
\name{bic_to_lk}
\alias{bic_to_lk}
\title{Convert Bayesian Information Criterion (BIC) to likelihood (lk) value}
\usage{
bic_to_lk(bic, k, n)
}
\arguments{
\item{bic}{Bayesian Information Criterion (BIC) value}

\item{k}{number of parameters estimated}

\item{n}{number of observations}
}
\description{
Converts Bayesian Information Criterion (BIC) to likelihood (lk) value
}
\examples{
bic_to_lk()
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
