% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vec.R
\name{check_vec}
\alias{check_vec}
\title{check_vec}
\usage{
check_vec(
  x,
  y,
  nm_fcn = NULL,
  test_1 = function(x) {
     is.numeric(x)
 },
  test_2 = function(x, y) {
     length(x) == length(y)
 }
)
}
\arguments{
\item{nm_fcn}{optional name of the function the check is being run in}

\item{test_1}{logical test function 1}

\item{test_2}{logical test function 2}

\item{nm_x}{name of the object to check}

\item{nm_y}{name of the reference object}
}
\description{
Run checks on input vector
}
\details{
Used inside bamExtras functions to check if input vectors are numeric and are the same length as a reference object nm_y. Often vectors need to be numeric and the same length as a vector of ages, and the function shouldn't run unless these criteria are met.
}
\examples{
aser <- rdat_BlackSeaBass$a.series
M <- aser$M
age <- aser$age
check_vec("M","age") # Should invisibly return TRUE

# Generate an error
M <- M[-1]
check_vec("M","age",nm_fcn="pop_demo")

}
\author{
Nikolai ?warn
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
