% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_trans.r
\name{color_trans}
\alias{color_trans}
\title{Geometric mean: logarithm form}
\usage{
color_trans(hexcolor, alpha = 0.25)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Simple geometric mean function using only is.finite values of log(x).
}
\examples{
\dontrun{
n <- 5
mycol <- rainbow(n)
mycol_tr <- color_trans(mycol,alpha=0.5)
mycol2 <- rgb(1,seq(0,1,length=n),0)
mycol2_tr <- color_trans(mycol2)

par(cex=2,pch=16)
plot(1:n,rep(1,n),col=mycol,ylim=c(0,5))
points(1:n,rep(2,n),col=mycol_tr)
points(1:n,rep(3,n),col=mycol2)
points(1:n,rep(4,n),col=mycol2_tr)
}
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
