% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_combine.r
\name{comp_combine}
\alias{comp_combine}
\title{Complete and combine age or length comps}
\usage{
comp_combine(
  comp_list,
  FUN = function(x) {
     sum(x, na.rm = TRUE)
 },
  scale_rows = TRUE
)
}
\arguments{
\item{comp_list}{list of composition matrices, where column names are bin values (e.g. length or age), and rows are unique observations (e.g. years)}

\item{FUN}{function (unquoted name) to apply to analogous cells of comp matrices. Default is to sum values.}

\item{scale_rows}{logical. Should rows of resulting matrix be scaled to sum to 1? Defaults to TRUE.}
}
\description{
Complete and combine age or length comps
}
\examples{
\dontrun{
# Combine age comp matrices for Red Porgy
cm <- rdat_RedPorgy$comp.mats
cma <- cm[grepl("^acomp.*ob$",names(cm))]
comp_combine(cma)
}

}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
