% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_complete_bam.r
\name{comp_complete_bam}
\alias{comp_complete_bam}
\title{Complete a length or age composition data frame so that all have desired dimensions. This is a modified version of comp_complete. The two functions should probably be combined with options to choose between sub-functions.}
\usage{
comp_complete_bam(
  D_cmp,
  n_colnames,
  val_rownames,
  val_colnames,
  minusGroup = FALSE,
  plusGroup = FALSE
)
}
\arguments{
\item{D_cmp:}{comp matrices, where column names are sample size columns (e.g. n.fish, n.trips) as well as bin values (e.g. length or age), and rows are unique observations (e.g. years)}

\item{n_colnames:}{column names for sample size columns}

\item{val_rownames:}{row names you want to have included in each comp matrix}

\item{val_colnames:}{column names you want to have included in each comp matrix (excluding sample size columns)}

\item{minusGroup:}{logical. When truncating bin range, should valus below the smallest bin be summed into the smallest bin? The default is to simply truncate.}

\item{plusGroup:}{logical. When truncating bin range, should valus above the largest bin be summed into the largest bin? The default is to simply truncate.}
}
\description{
Complete a length or age composition data frame so that all have desired dimensions. This is a modified version of comp_complete. The two functions should probably be combined with options to choose between sub-functions.
}
\examples{
comp_complete_bam()
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
