% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2csv.r
\name{cor2csv}
\alias{cor2csv}
\title{Convert BAM .cor file to csv}
\usage{
cor2csv(cor_file, rdat = NULL)
}
\arguments{
\item{cor_file}{cor file path}

\item{rdat}{BAM output rdat (list) object read in with dget(). Optional, but the function can do more if you supply the rdat}
}
\description{
Convert BAM .cor file to .csv file. Also invisibly returns the cor matrix.
}
\examples{
\dontrun{
}

}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
