% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDecay.r
\name{expDecay}
\alias{expDecay}
\title{Exponential decay}
\usage{
expDecay(age, Z, N0 = 1, plus.group = FALSE)
}
\arguments{
\item{age}{ages}

\item{Z}{instantaneous mortality rate}

\item{N0}{number of individuals at age zero at the beginning of the year}

\item{plus.group}{Should the function include a plus group? logical}
}
\description{
Compute numbers of individuals alive at time in a population experiencing exponential decay, allowing age-varying Z. Can be used to compute numbers at age at the beginning of the year.
}
\examples{
\dontrun{
expDecay(age=1:10,Z=0.2,N0=100)
expDecay(age=1:10,Z=0.2,N0=100,plus.group = TRUE)
expDecay(age=c(2,5,10),Z=0.2,N0=100,plus.group = TRUE)
}
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
