% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs2tex.R
\name{figs2tex}
\alias{figs2tex}
\title{Loop through figure files (e.g. pdf) in a folder and generate LaTeX code to make a beamer slide for each figure.
This function is developed based on the folder structure generated by the fishgraph package}
\usage{
figs2tex(
  dir1,
  folders,
  dir2 = NULL,
  ext = "pdf",
  label1 = NULL,
  label2 = NULL,
  height = NULL,
  dir_prefix = NULL,
  file_root = NULL,
  hypertargetText = ""
)
}
\arguments{
\item{dir1}{directory containing subdirectories (folders) of figures (e.g. spp directory produced by fishgraph)}

\item{folders}{names of folders of interest within dir1}

\item{dir2}{optional second directory, analogous to dir1. If provided, the function will generate text code to produce two column slides for comparing analogous figures (e.g. from two runs of the same model)}

\item{ext}{file extension of figues in subdirectories}

\item{label1}{text label for slides associated with dir1}

\item{label2}{text label for slides associated with dir1}

\item{height}{figure height in cm, passed to the height argument of the latex command includegraphics}

\item{dir_prefix}{annoying prefix on folder names to be removed with gsub (e.g. figs)}

\item{file_root}{root names (i.e. without ext) of figures to plot}

\item{hypertargetText}{Character vector of names for creating hyperlinks in slides (I'll be more specific when I remember what this does!)}
}
\description{
Loop through figure files (e.g. pdf) in a folder and generate LaTeX code to make a beamer slide for each figure.
This function is developed based on the folder structure generated by the fishgraph package
}
\examples{
\dontrun{
}
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
