% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MCBE.r
\name{filter_MCBE}
\alias{filter_MCBE}
\title{Identify which runs should be filtered out of results of MCBE uncertainty analysis}
\usage{
filter_MCBE(
  sim_summary,
  filter_info = list(gradient.max = c(0, 0.01), Fref = c(0, 5), F.Fref = c(0, 5),
    R.sigma.par = c(0.2, 1), R0_prob = c(0.005, 0.995), avoid_bounds = TRUE),
  nm_rp = list(parms = c(Fref = "Fmsy", F.Fref = "Fend.Fmsy.mean"))
)
}
\arguments{
\item{sim_summary}{Output object from summarize_MCBE}

\item{filter_info}{list of ranges used to filter out MCBE results. set to NULL to retain all results.}
}
\value{
This function returns a list including: 1. \code{sim_pass} is a
numeric vector indicating which simulation runs should be retained after filtering.
2. \code{test_vals} is a data frame of the values used to conduct filtering for all runs.
3. \code{test_logical} is a logical matrix indicating the results of each
filtering test.
}
\description{
Identify which runs should be filtered out of results of MCBE uncertainty analysis
}
\examples{
\dontrun{
# Run MCBE, writing files to dir_bam_sim
run_MCBE("GrayTriggerfish", dir_bam_sim="sim_GrTr")
# Summarize results of MCBE and assign to object
ss_GrTr <- summarize_MCBE(dir_bam_sim="sim_GrTr")
sim2keep <- filter_MCBE(ss_GrTr,nm_rp = list("spr.brps"=c("Fref"="F30","F.Fref"="Fend.F30.mean")))
# Plot MCBE results
}
}
\author{
Kyle Shertzer, Erik Williams, and Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
