% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean2.r
\name{geomean2}
\alias{geomean2}
\title{Geometric mean: product form}
\usage{
geomean2(x, na.rm = TRUE, all.na.result = NA)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{as in base R functions "a logical value indicating whether NA values should be stripped before the computation proceeds."}

\item{all.na.result}{Result to return if all values are NA. For some reason prod(numeric(0)) = 1, so the function will return 1 even if all x are NA.}
}
\description{
Simple geometric mean function.
The logarithm form (geomean) works somewhat better but can only handle non-zero positive values.
}
\examples{
x <- c(1,2,10)
geomean(x)
geomean2(x)
mean(x)
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
