% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msy.r
\name{get_msy}
\alias{get_msy}
\title{get_msy}
\usage{
get_msy(
  rdat,
  maxF = NULL,
  nFopt = NULL,
  nFres = 100,
  SPR_props = c(0.3, 0.4, 0.5),
  nages = NULL,
  sel_wgted_L = NULL,
  sel_wgted_D = NULL,
  M = NULL,
  reprod = NULL,
  spawn_time_frac = NULL,
  R0 = NULL,
  steep = NULL,
  BiasCor = NULL,
  SR_switch = NULL,
  wgt_mt = NULL,
  wgt_klb = NULL,
  wgt_klb_L = NULL,
  wgt_klb_D = NULL,
  make_plot = FALSE,
  plot_digits = 3
)
}
\arguments{
\item{rdat}{BAM output rdat file object}

\item{maxF}{Maximum fishing mortality rate. numeric vector}

\item{nFopt}{length of F vector for grid search optimization}

\item{nFres}{length of F vector for filtering grid search results to limit output size}

\item{SPR_props}{proportions of SPR to compute reference points at. numeric vector}

\item{nages}{Number of ages in population model. numeric vector}

\item{sel_wgted_L}{effort-weighted, recent selectivities toward landings. numeric vector of length \code{nages}}

\item{sel_wgted_D}{effort-weighted, recent selectivities toward discards. numeric vector of length \code{nages}}

\item{M}{Natural mortality rate. numeric vector of length \code{nages}}

\item{reprod}{Reproductive potential at-age vector. numeric vector of length \code{nages}}

\item{spawn_time_frac}{time of year of peak spawning, as a fraction of the year. numeric}

\item{R0}{Beverton-Holt R0 parameter (virgin recruitment). Numbers of fish at first age (often age-0 or age-1). numeric vector}

\item{steep}{Beverton-Holt steepness parameter. numeric vector}

\item{BiasCor}{bias correction. numeric vector}

\item{SR_switch}{1 for Beverton-Holt, 2 for Ricker, 3 for none (average recruitment). integer}

\item{wgt_mt}{Weight-at-age in metric tons for fish in the population. numeric vector of length \code{nages}}

\item{wgt_klb}{Weight-at-age in thousand pounds for fish in the population. numeric vector of length \code{nages}}

\item{wgt_klb_L}{Weight-at-age in thousand pounds for fish in the landings. numeric vector of length \code{nages}}

\item{wgt_klb_D}{Weight-at-age in thousand pounds for fish in the discards. numeric vector of length \code{nages}}

\item{make_plot}{Indicate whether or not to plot results. logical}

\item{plot_digits}{number of significant digits to show in plot. numeric}
}
\description{
Calculations of stock dependent quantities over a range of fishing mortalities (F). This is code from the Beaufort Assessment Model converted from ADMB to R.
}
\examples{
\dontrun{
rdat <- rdat_VermilionSnapper
res <- get_msy(rdat,make_plot=TRUE)
}
}
\author{
Erik Williams, Kyle Shertzer, and Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
