% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_pts.r
\name{get_ref_pts}
\alias{get_ref_pts}
\title{get_ref_pts}
\usage{
get_ref_pts(
  rdat = NULL,
  maxF = NULL,
  nFopt = 10001,
  nFres = 101,
  SPR_props = c(0.3, 0.4, 0.5),
  nages = NULL,
  sel_wgted_L = NULL,
  sel_wgted_D = NULL,
  M = NULL,
  reprod = NULL,
  spawn_time_frac = NULL,
  R0 = NULL,
  steep = NULL,
  BiasCor = NULL,
  SR_switch = NULL,
  R0_2 = NULL,
  steep_2 = NULL,
  BiasCor_2 = NULL,
  SR_switch_2 = "none",
  rec_mean = NULL,
  wgt_mt = NULL,
  wgt_klb = NULL,
  wgt_klb_L = NULL,
  wgt_klb_D = NULL,
  make_plot = FALSE,
  plot_digits = 4
)
}
\arguments{
\item{rdat}{BAM output rdat file object}

\item{maxF}{Maximum fishing mortality rate. numeric vector}

\item{nFopt}{length of F vector for grid search optimization}

\item{nFres}{length of F vector for filtering grid search results to limit output size}

\item{SPR_props}{proportions of SPR to compute reference points at. numeric vector}

\item{nages}{Number of ages in population model. numeric vector}

\item{sel_wgted_L}{effort-weighted, recent selectivities toward landings. numeric vector of length \code{nages}}

\item{sel_wgted_D}{effort-weighted, recent selectivities toward discards. numeric vector of length \code{nages}}

\item{M}{Natural mortality rate. numeric vector of length \code{nages}}

\item{reprod}{Reproductive potential at-age vector. numeric vector of length \code{nages}}

\item{spawn_time_frac}{time of year of peak spawning, as a fraction of the year. numeric}

\item{R0}{Beverton-Holt R0 parameter (virgin recruitment) used in equilibrium calculations. Numbers of fish at first age (often age-0 or age-1). numeric vector}

\item{steep}{Beverton-Holt steepness parameter used in equilibrium calculations. numeric vector}

\item{BiasCor}{bias correction used in equilibrium calculations. numeric vector}

\item{SR_switch}{1 for Beverton-Holt, 2 for Ricker, 3 for none (average recruitment) used in equilibrium calculations. integer}

\item{R0_2}{Beverton-Holt R0 parameter (virgin recruitment) used in non-equilibrium calculations. Numbers of fish at first age (often age-0 or age-1). numeric vector}

\item{steep_2}{Beverton-Holt steepness parameter used in non-equilibrium calculations. numeric vector}

\item{BiasCor_2}{bias correction used in non-equilibrium calculations. numeric vector}

\item{SR_switch_2}{1 for Beverton-Holt, 2 for Ricker, 3 for none (average recruitment) used in non-equilibrium calculations. integer}

\item{wgt_mt}{Weight-at-age in metric tons for fish in the population. numeric vector of length \code{nages}}

\item{wgt_klb}{Weight-at-age in thousand pounds for fish in the population. numeric vector of length \code{nages}}

\item{wgt_klb_L}{Weight-at-age in thousand pounds for fish in the landings. numeric vector of length \code{nages}}

\item{wgt_klb_D}{Weight-at-age in thousand pounds for fish in the discards. numeric vector of length \code{nages}}

\item{make_plot}{Indicate whether or not to plot results. logical}

\item{plot_digits}{number of significant digits to show in plot. numeric}
}
\value{
Invisibly returns a list including data generated by the equilibrium
and non-equilibrium calculations (\code{data_eq} and \code{data_neq}), as
well as MSY and SPR-based reference points (\code{ref_pts_msy} and \code{ref_pts_SPR}),
computed with equilibrium and non-equilibrium methods, respectively.
}
\description{
Calculations of stock dependent quantities over a range of fishing mortalities (F).
This is code based on code from the Beaufort Assessment Model converted from ADMB to R,
for both bam functions \code{get_msy} and \code{get_per_recruit_stuff}.
}
\details{
This functions computes msy-based reference points with equilibrium methods,
and computes SPR-based reference points using non-equilibrium methods. Both sets
of calculations use the same function, \code{sr_eq}, to calculate recruitment
at each level of F, but separate sets of arguments are passed to each call.
By default, \code{SR_switch_2="none"}, setting recruitment to \code{BiasCor_2*R0_2}.
To set recruitment equal to a fixed value (e.g. empirical mean recruitment from a set of years),
set \code{R0_2} equal to that value, and set \code{BiasCor_2} equal to 1.
}
\examples{
\dontrun{
rdat <- rdat_VermilionSnapper
res <- get_ref_pts(rdat,make_plot=TRUE)
# Using average recruitment in the non-equilibrium calculations, as in SEDAR 55
rec_mean <- mean(rdat$t.series[paste(1976:2013),"recruits"])
res2 <- get_ref_pts(rdat,R0_2=rec_mean,BiasCor_2=1,make_plot=TRUE)
}
}
\author{
Erik Williams, Kyle Shertzer, and Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
