% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tpl_nm.r
\name{get_tpl_nm}
\alias{get_tpl_nm}
\title{Get the names of all objects named in the BAM tpl file, with object classes in \code{tpl_object_classes}.}
\usage{
get_tpl_nm(
  tpl_obj = NULL,
  tpl_file = NULL,
  tpl_object_classes = c("init_int", "number", "init_number", "init_bounded_number",
    "const double", "vector", "init_vector", "init_bounded_vector", "init_ivector",
    "init_bounded_dev_vector", "matrix", "init_matrix")
)
}
\arguments{
\item{tpl_obj}{tpl file read in as a character vector with readLines(con=tpl_file)}

\item{tpl_file}{tpl file path}

\item{tpl_obj_classes}{Character vector of names of tpl object classes. I think I listed all of the classes used in BAM, but I might have missed something.}
}
\description{
Get the names of all objects named in the BAM tpl file, with object classes in \code{tpl_object_classes}.
}
\examples{
get_tpl_nm(tpl_AtlanticMenhaden)
get_tpl_nm(tpl_BlackSeaBass)

}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
