% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputData_Tilefish}
\alias{inputData_Tilefish}
\title{Raw input data: South Atlantic Tilefish (SEDAR 66)}
\format{
An object of class \code{"list"}
\describe{
\item{acomp.cH}{Age composition data for the commercial handline fleet}
\item{acomp.cL}{Age composition data for the commercial longline fleet}
\item{acomp.mm}{Age composition data for the SCDNR MARMAP longline Survey}
\item{BAMDataBAMDataLikWeights}{A table of values representing inputs to the likelihood weights section of the BAM dat file. Note that these values may not have been directly used to modify the dat file.}
\item{BAMDataParPoints}{A table of values representing inputs to point parameters (e.g. BH steepness) in the parameters section of the BAM dat file. Note that these values may not have been directly used to modify the dat file.}
\item{BAMDataBAMDataParVectors}{A table of values representing inputs to dev vectors (e.g. set_log_rec_dev) in the BAM dat file. Note that these values may not have been directly used to modify the dat file.}
\item{Indices}{Time series of abundance indices.}
\item{Landings}{Time series of landings and landings cvs by fleet.}
\item{lcomp.cH}{Length composition data for the commercial handline fleet}
\item{lcomp.cL}{Length composition data for the commercial longline fleet}
\item{lcomp.rA}{Length composition data for the recreational (All sources) fleet}
\item{LifeHistoryAtAge}{Life history data by age}
\item{LifeHistoryPointValues}{Other miscellaneous point values used in the assessment, such as estimates of constant natural mortality and meristics equation parameters.}
}
}
\source{
Data providers submitted data to NMFS Beaufort which was then compiled into spreadsheets.
}
\usage{
inputData_Tilefish
}
\description{
Raw input data: South Atlantic Tilefish (SEDAR 66)
}
\keyword{datasets}
