% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2F.r
\name{L2F}
\alias{L2F}
\title{L2F}
\usage{
L2F(L_tar, F_est, sel_L, sel_Z, M, N_a, wgt_L, control = list(iter.max = 500))
}
\arguments{
\item{L_tar}{target landings value to match to the target fishing mortality value, F_tar. Must be in the same units as wgt_L.}

\item{F_est}{initial guesses for the F that matches L_tar}

\item{sel_L}{selectivity for landings}

\item{sel_Z}{selectivity for total mortality}

\item{M}{natural mortality rate, at age}

\item{N_a}{vector of numbers of fish, at age, in population}

\item{wgt_L}{vector of weight of fish, at age, in the landings}

\item{control}{A list of control parameters passed to \code{\link[stats]{nlminb}}}
}
\description{
Approximate level of fishing mortality that achieves a given level of landings.
This function is modified from code that was part of the BAM projections. The
current function is intended for use prior to running run_proj to compute the
F that matches specific values of landings. This is particularly useful if
landings estimates corresponding to a year or years between the terminal year of the
current stock assessment and the start year of new management is available.
}
\examples{
\dontrun{
spp <- rdat_Tilefish
endyr <- paste(spp$parms$endyr)
# Compute landings in weight for the terminal year
L_endyr_klb <- L_calc(F_a = spp$F.age[endyr,], Z_a = spp$Z.age[endyr,],
N_a = spp$N.age[endyr,], wgt_L = spp$a.series$gutwgt.wgted.L.klb)
L2F(L_tar=L_endyr_klb, F_est=sum(spp$F.age[endyr,]),sel_L = spp$sel.age$sel.v.wgted.L,
sel_Z = spp$sel.age$sel.v.wgted.tot,M = spp$a.series$M, N_a = spp$N.age[endyr,], wgt_L = spp$a.series$gutwgt.wgted.L.klb)
}
}
\author{
Kyle Shertzer and Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
