% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_calc.r
\name{L_calc}
\alias{L_calc}
\title{L_calc}
\usage{
L_calc(F, Z, N, wgt_L = NULL)
}
\arguments{
\item{F}{vector of fishing mortality rate, at age}

\item{Z}{vector of total mortality rate at age including dead discards}

\item{N}{vector of numbers of fish, at age, in population}

\item{wgt_L}{vector of weight of fish, at age, in the landings. units of the result will be in units of wgt_L (e.g. metric tons, 1000 lb)}
}
\description{
Compute landings in numbers or weight with Baranov equation. Set wgt_L = NULL to compute numbers instead of weight. Used in run_proj.
}
\examples{
\dontrun{
spp <- rdat_Tilefish
endyr <- paste(spp$parms$endyr)

# Compute landings in numbers for the terminal year
sum(L_calc(F = spp$F.age[endyr,], Z = spp$Z.age[endyr,], N = spp$N.age[endyr,]))
# The value should be very close to this:
sum(spp$CLD.est.mats$Ln.total[endyr,])
# Compute landings in weight for the terminal year
sum(L_calc(F = spp$F.age[endyr,], Z = spp$Z.age[endyr,], N = spp$N.age[endyr,], wgt_L = spp$a.series$gutwgt.wgted.L.klb))
# The value should be very close to this:
sum(spp$CLD.est.mats$Lw.total[endyr,])
}
}
\author{
Kyle Shertzer and Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
