% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgs.r
\name{lgs}
\alias{lgs}
\title{lgs}
\usage{
lgs(x, a = 1, b = 0, par = NULL, type = "slope_inflection")
}
\arguments{
\item{x}{numeric vector}

\item{a}{slope parameter}

\item{b}{point of inflection (e.g. a50)}

\item{par}{named vector of parameters a and b. If this is specified it will}

\item{type}{which form of the logistic function to use: "slope_intercept" for
the slope-intercept form, "slope_inflection" for the slope-point-of-inflection form.
For both forms, a is the slope parameter.
override specifications of individual parameters}
}
\description{
logistic function
}
\examples{
\dontrun{
x <- 1:40
y <- lgs(x=x, a=1, b=5)
plot(x,y)

# Alternative way to specify parameters
y <- lgs(x=x, par=c(x=x, a=1, b=5))
plot(x,y)
}
}
\author{
Nikolai Klibansky
}
