% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm_cv.r
\name{lnorm_cv}
\alias{lnorm_cv}
\title{Coefficient of variation (cv) of lognormal distribution}
\usage{
lnorm_cv(sigma)
}
\arguments{
\item{sigma}{standard deviation of log(x)}
}
\description{
Coefficient of variation (cv) of lognormal distribution x, given the standard deviation (sigma) of log(x). Bolker 2008 page 137
}
\examples{
\dontrun{
set.seed(1234)
x <- rnorm(1000,mean=100,sd=3)
s <- sd(log(x))
lnorm_cv(sigma=s)
}

}
\references{
Bolker, B. M. 2008. Ecological Models and Data in R. Princeton University Press, Princeton, NJ. Page 137
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
