% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm_cv_to_sd.r
\name{lnorm_cv_to_sd}
\alias{lnorm_cv_to_sd}
\title{Convert lognormal cv to lognormal sd}
\usage{
lnorm_cv_to_sd(cv)
}
\arguments{
\item{cv}{lognormal cv}
}
\description{
Convert lognormal cv to lognormal sd
}
\examples{
set.seed(1234)
x <- rnorm(1000,mean=100,sd=3)
m <- mean(log(x))
s <- sd(log(x))
lnorm_sd(mu=m,sigma=s)
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
