% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm_sd.r
\name{lnorm_sd}
\alias{lnorm_sd}
\title{Standard deviation of lognormal distribution}
\usage{
lnorm_sd(mu, sigma)
}
\arguments{
\item{mu}{mean of log(x)}

\item{sigma}{standard deviation of log(x)}
}
\description{
Standard deviation of lognormal distribution x, given the mean (mu) and standard deviation (sigma) of log(x). Bolker 2008 page 137
}
\examples{
set.seed(1234)
x <- rnorm(1000,mean=100,sd=3)
m <- mean(log(x))
s <- sd(log(x))
lnorm_sd(mu=m,sigma=s)
}
\references{
Bolker, B. M. 2008. Ecological Models and Data in R. Princeton University Press, Princeton, NJ. Page 137
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
