% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm_vector_boot.r
\name{lnorm_vector_boot}
\alias{lnorm_vector_boot}
\title{Generating bootstrap vectors of with lognormal error}
\usage{
lnorm_vector_boot(x, cv, bootN = 1, standardize = FALSE, digits = NULL)
}
\arguments{
\item{x}{observed vector}

\item{cv}{observed cv for x}

\item{bootN}{number of bootstrap replicates. Allows multiple bootstrapped vectors to be calculated without looping}

\item{standardize}{divide resulting vectors by their own mean so that they have a mean of 1}

\item{digits}{Number of significant digits to round output to. If NULL, output is not rounded.}
}
\description{
This function was designed to be used with time series of indices of abundance or removals. Based on code from previous BAM
Monte-Carlo Bootstrap code
}
\examples{
\dontrun{
rdat <- rdat_VermilionSnapper
year <- rdat$t.series$year
U <- setNames(rdat$t.series$U.cHL.ob,year)
cv_U <- setNames(rdat$t.series$cv.U.cHL,year)
nsim <- 100
U_boot <- lnorm_vector_boot(U,cv_U,nsim)
dimnames(U_boot) <- list("year"=year,"sim"=sprintf(paste("\%0",nchar(nsim),".0f",sep=""),1:nsim))
plot_boot_vec(t(U_boot))
points(as.numeric(year),U,type="o",col="blue")
}
}
\author{
Nikolai Klibansky, Kyle Shertzer, Erik Williams
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
