% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_exponential.R
\name{logistic_exponential}
\alias{logistic_exponential}
\title{logistic-exponential function
logistic-exponential selectivity function. The function has 4 parameters.
When fitting in bam, one of the parameters should be fixed.
Note that if the minimim value of \code{ages} is equal to the value of \code{joint},
then \code{A50} and \code{slope} have no effect on the result.}
\usage{
logistic_exponential(ages, A50, slope, sigma, joint)
}
\arguments{
\item{ages}{vector of ages}

\item{A50}{age at 50\% selection (ascending limb)}

\item{slope}{rate of increase}

\item{sigma}{controls rate of descent (descending limb)}

\item{joint}{age to join curves}
}
\description{
logistic-exponential function
logistic-exponential selectivity function. The function has 4 parameters.
When fitting in bam, one of the parameters should be fixed.
Note that if the minimim value of \code{ages} is equal to the value of \code{joint},
then \code{A50} and \code{slope} have no effect on the result.
}
\examples{

rdat <- rdat_GrayTriggerfish
fp <- rdat$parm.cons[8,grepl("rHB\\\\.D$",names(rdat$parm.cons))]
logistic_exponential(ages=rdat$a.series$age,
A50 = fp$selpar.L50.rHB.D,
slope = fp$selpar.slope.rHB.D,
sigma = fp$selpar.sigma.rHB.D,
joint = fp$selpar.afull.rHB.D)
}
\author{
Erik Williams, Kyle Shertzer, and Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
