% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.r
\name{logit}
\alias{logit}
\title{logit}
\usage{
logit(x)
}
\arguments{
\item{x}{numeric vector (e.g. probabilities, 0 < x < 1)}
}
\description{
simple logit function.
}
\note{
This is really the same as \code{qlogis()} but it's nice to be able to just call \code{logit()}.
Returns Inf when x = 0 or x = 1 and NaN for x < 0 or x > 1.
}
\examples{
\dontrun{
x <- 1:20
y <- bamExtras::lgs(x=x)
z <- bamExtras::logit(x=y)
z2 <- stats::qlogis(y)
plot(x,z)
points(x,z2,type="l",col="blue")
}
}
\author{
Nikolai Klibansky
}
