% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_calc.r
\name{M_calc}
\alias{M_calc}
\title{Compute constant natural mortality estimates from life history parameters}
\usage{
M_calc(type = "all", tmax, Linf, K)
}
\arguments{
\item{type}{Type of M calculation to use. If type="all" the function tries all calculations
and returns a vector of all results. NA values indicate that a particular calculation couldn't
be computed, probably because a necessary life history parameter was not provided.
Possible values include four calculations specified in Hoenig (1983; Table 1;
"Hng_Mollusks", "Hng_Fish", "Hng_Cetaceans", "Hng_All") and six calculations
from Then et al. (2014; Table 3; "Thn_1partmax", "Thn_Hnglm", "Thn_Hngnls",
"Thn_1parK", "Thn_2parK", "Thn_Plynls").}

\item{Linf}{length infinity in cm (von Bertalanffy growth equation parameter)}

\item{K}{growth coefficient (von Bertalanffy growth equation parameter)}

\item{t0}{age (time) at length zero (von Bertalanffy growth equation parameter)}
}
\description{
Compute constant natural mortality estimates from life history parameters
}
\examples{
\dontrun{
# Return all possible estimates based on tmax
M_calc(tmax=11)
# Return all possible estimates based on tmax or K
M_calc(tmax=11,K=0.173)
# Return all possible estimates based on tmax, K, or Linf (in centimeters)
M_calc(tmax=11,Linf=50.2,K=0.173)
# Return just the Hoenig_nls estimate published in Then et al. (2014) based on tmax
M_calc(type="Thn_Hngnls",tmax=11)
}
}
\references{
Hoenig, J. M. 1983. Empirical use of longevity data to estimate mortality-rates. Fishery Bulletin 81:898-903.

Then, A. Y., J. M. Hoenig, N. G. Hall, and D. A. Hewitt. 2014. Evaluating the predictive performance of empirical estimators of natural mortality rate using information on over 200 fish species. ICES Journal of Marine Science.
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
