% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_Lorenzen.R
\name{M_Lorenzen}
\alias{M_Lorenzen}
\title{Compute age-varying natural mortality (M) based on Lorenzen et al (1996) and scale based on constant M}
\usage{
M_Lorenzen(
  a,
  Linf,
  K,
  t0,
  par_wl_a = 1,
  par_wl_b = 3,
  aP = 0,
  M_constant = NULL,
  aMin = 0,
  par_a = 3.69,
  par_b = -0.305
)
}
\arguments{
\item{a}{ages at which to compute M}

\item{Linf}{length infinity}

\item{K}{growth coefficient}

\item{t0}{age (time) at length zero}

\item{par_wl_a}{a parameter from weight~length conversion equation weight = par_wl_a*length^par_wl_b}

\item{par_wl_b}{b parameter from weight~length conversion equation weight = par_wl_a*length^par_wl_b}

\item{aP}{proportion of age (value between 0 and 1) at which to compute length (e.g. aP=0.5 to compute length at midyear)}

\item{M_constant}{constant M value used to scale age-varying M. Defaults to NULL. M will be scaled if a numeric value of M_constant is supplied.}

\item{aMin}{minimum age to include in computation of scaling factor when scaling M}

\item{par_a}{a parameter for Lorenzen equation par_a*W^par_b. Defaults to value from Lorenzen et al. (1996; Table 1, ocean system) par_a=3.69.}

\item{par_b}{b parameter for Lorenzen equation par_a*W^par_b. Defaults to value from Lorenzen et al. (1996; Table 1, ocean system) par_b=-0.305.}
}
\description{
Compute age-varying natural mortality (M) based on Lorenzen et al (1996) and scale based on constant M
}
\note{
Input parameters for von Bertalanffy growth equation and weight~length equation should have units that result in weight in grams (g), which is what the Lorenzen equation is based on.
}
\examples{
\dontrun{
# M for Black Sea Bass
M_Lorenzen(a=0:11,Linf = 502, K = 0.173, t0 = -0.97, M_constant=0.38, aMin=2, par_wl_a=5.02e-05, par_wl_b=2.77)
}

}
\references{
Lorenzen, K. 1996. The relationship between body weight and natural mortality in juvenile and adult fish: a comparison of natural ecosystems and aquaculture. Journal of Fish Biology 49:627-642.
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
