% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_noNA.r
\name{mean_noNA}
\alias{mean_noNA}
\title{Computes mean excluding NA values}
\usage{
mean_noNA(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Version of mean() where default na.rm=TRUE
}
\examples{
x <- c(1,2,10,NA)
mean(x)
mean(x,na.rm=TRUE)
mean_noNA(x)
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
