% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_LFS.R
\name{ML_LFS}
\alias{ML_LFS}
\title{Compute mean lengths by year above a certain minimum size (e.g. length at full selection).}
\usage{
ML_LFS(CAL, minL = 0)
}
\arguments{
\item{CAL}{catch-at-length matrix, in numbers (rows=years, columns=length bins)}

\item{minL}{minimum length. Allows you to truncate the matrix to values above this value}
}
\description{
Compute mean lengths by year above a certain minimum size (e.g. length at full selection).
}
\examples{
\dontrun{

}
}
\author{
Nikolai Klibansky
}
\keyword{DLMtool}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
