% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse_calc.r
\name{mse_calc}
\alias{mse_calc}
\title{Compute mean squared errors (MSE) for indices}
\usage{
mse_calc(x)
}
\arguments{
\item{x}{a BAM rdat object, i.e. x <- dget(rdat)}
}
\description{
Compute mean squared errors (MSE) for indices
}
\examples{
\dontrun{
mse_calc(rdat_BlackSeaBass)
}

}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
