% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boot_density.R
\name{plot_boot_density}
\alias{plot_boot_density}
\title{Make density plots of bootstrapped parameters and add quantiles as vertical lines.}
\usage{
plot_boot_density(
  data,
  x_name,
  density_args = list(from = NULL, to = NULL),
  plot_args = list(xlab = NULL, ylab = "density", main = "", type = "l"),
  quantile_draw = FALSE,
  quantile_args = list(probs = c(0.05, 0.5, 0.95)),
  quantile_abline_args = list(lty = c(3, 2, 3), lwd = c(2, 2, 2), lend = "butt"),
  col_shade = "gray50",
  add_text_nsim = FALSE
)
}
\arguments{
\item{data}{a numeric matrix or data frame that can be coerced to a numeric matrix where rows are runs and columns represent parameters or variables.}

\item{x_name}{column name of a variable x to plot}

\item{density_args}{arguments passed to \code{stats::density} \code{\link[stats]{density}} (excluding x)}

\item{plot_args}{arguments passed to \code{base::plot} \code{\link[base]{plot}} (excluding x and y)}

\item{quantile_draw}{logical. Should quantiles of x be added to the plot as vertical lines? (e.g. confidence limits and median)}

\item{quantile_args}{arguments passed to \code{stats::quantile} \code{\link[stats]{quantile}} for computing quantiles of x to add to plot as vertical lines}

\item{quantile_abline_args}{arguments to pass to \code{graphics::abline} \code{\link[graphics]{abline}} when plotting quantiles as vertical lines (excluding v)}

\item{col_shade}{color for shading density region}

\item{add_text_nsim}{logical. Should the number of simulation runs (nsim) in \code{data} be added as text to the plots?}
}
\description{
Make density plots of bootstrapped parameters and add quantiles as vertical lines.
}
\examples{
\dontrun{
nsim <- 100
data <- data.frame("M"=runif(nsim,0.19, 2.1),"lwb"=rnorm(nsim,mean=3))
par(mfrow=c(2,2))
plot_boot_density(data,"M")
plot_boot_density(data,"M",add_text_nsim = TRUE)
plot_boot_density(data,"lwb")
plot_boot_density(data,"lwb",add_text_nsim = TRUE,quantile_draw=TRUE,col_shade=rgb(0,0,1,0.5))
}
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
