% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boot_phase.R
\name{plot_boot_phase}
\alias{plot_boot_phase}
\title{Phase plots for bootstrapped data}
\usage{
plot_boot_phase(
  x,
  y,
  xref = NULL,
  yref = NULL,
  col_point = rgb(0, 0, 0, 0.2),
  col_quant = "green4",
  col_ref = "black",
  col_text = "blue",
  probs = c(0.05, 0.95),
  text_x_adj = c(0, 0, 0, 0),
  text_y_adj = c(0, 0, 0, 0),
  ...
)
}
\arguments{
\item{x}{x values from bootstrapping. numeric vector}

\item{y}{y values from bootstrapping. numeric vector}

\item{xref}{x reference value (e.g. base run value).}

\item{yref}{y reference value (e.g. base run value).}

\item{col_point}{color for (x,y) points}

\item{col_quant}{color for quantile lines}

\item{col_ref}{color for (xref,yref) points}

\item{col_text}{color for percentages plotted in each quadrant}

\item{probs}{quantile probabilities to compute length of error bars.}

\item{text_x_adj}{text (percentages in each quadrant) adjustment to x}

\item{text_y_adj}{text (percentages in each quadrant) adjustment to y}

\item{...}{other parameters to pass to \code{plot}}
}
\description{
Bivariate scatter plots, specially formatted for making phase plots for stock and fishery status results.
}
\examples{
nsim <- 1000
x <- rnorm(n=nsim,mean=1.1,sd=.2)
y <- 1*(1/x)+rnorm(n=nsim,mean=0.1,sd=.1)
par(mfrow=c(2,2))
plot_boot_phase(x,y)
plot_boot_phase(x,y,xref=1.2,yref=1.1)
plot_boot_phase(x,y,xref=1.2,yref=1.1,xlab="F/Fmsy",ylab="SSB/SSBmsy")
plot_boot_phase(x,y,xref=1.2,yref=1.1,xlab="F/Fmsy",ylab="SSB/SSBmsy",col_point=rainbow(nsim,alpha=0.3))
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
