% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boot_vec.R
\name{plot_boot_vec}
\alias{plot_boot_vec}
\title{Plot stochastic vectors}
\usage{
plot_boot_vec(
  data,
  CIpct = 95,
  runsToKeep = NULL,
  ref_x = NULL,
  ref_y = NULL,
  col_lines = rgb(0, 0, 0, 1),
  col_shade = NULL,
  add = FALSE,
  ...
)
}
\arguments{
\item{data}{a numeric matrix or data frame that can be coerced to a numeric matrix where rows are runs and columns represent a  numeric sequence (e.g. age, year). Note that the function would probably work for any numeric matrix, but it's intended for bootstrapped time- or age-series data.}

\item{CIpct}{percentage to define the range of the confidence intervals to plot}

\item{runsToKeep}{optional vector of integer values corresponding to the rows of \code{data} to include in the plot}

\item{ref_x, ref_y}{x and y vectors for adding a reference line to the plot (e.g. base run values)}

\item{col_lines}{color for lines depicting central tendency and confidence limits. Expects hex color}

\item{col_shade}{color for shading confidence bands}

\item{add}{logical. Should a new plot be drawn or should it be added to an existing plot?}

\item{...}{other arguments to pass to \code{matplot}}
}
\description{
Plot median values and confidence intervals (shaded region) for stochastic (e.g. bootstrapped) vectors.
}
\examples{
\dontrun{
bootData <- t(replicate(n=100,random_walk(mu=0.5)))
plot_boot_vec(bootData)
}
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
