% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_timeline.R
\name{plot_data_timeline}
\alias{plot_data_timeline}
\title{plot_data_timeline}
\usage{
plot_data_timeline(
  rdat,
  key_data_type_abb = c(L = "landings", D = "discards", U = "index", acomp = "age comp",
    lcomp = "length comp"),
  key_fleet_abb = c(cHL = "com handline", cOT = "com other", cLL = "com longline", rGN =
    "rec general", rHB = "rec headboat", rHB.D = "rec headboat discards", sTV =
    "SERFS trap video"),
  data_types = c("L", "D", "U", "acomp", "lcomp"),
  FUN_col = "rainbow",
  t.series.row.trunc = 1
)
}
\arguments{
\item{rdat}{(list) object read in with dget()}

\item{key_data_type_abb}{data type abbreviation key}

\item{key_fleet_abb}{fleet abbreviation key}

\item{data_types}{data types to include in plot}

\item{t.series.row.trunc}{truncate rows of t.series object by this value to
avoid plotting projection year or show what available data would look
like in retrospective runs}
}
\description{
Make a data timeline table from the BAM rdat output t.series object. Note: plot_data_timeline() uses the plot.matrix package
}
\author{
Kyle Shertzer and Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
