% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot results from MCBE parameter profile analysis}
\usage{
plot_profile(
  sim_summary,
  nm_par = "steep",
  nm_obj = c("parms", "parm.cons"),
  like_pattern = "^lk.",
  like_groups = c("L", "D", "U", "lenc", "agec", "SR"),
  like_shift = "min_par",
  like_exclude = c("lk.unwgt.data", "lk.Nage.init", "lk.fullF", "lk.Ftune", "lk.SRearly",
    "lk.SRend"),
  like_exclude_stacked = c("lk.total", "lk.unwgt.data"),
  like_plot_last = "lk.total",
  par_args = list(mfrow = c(2, 1), mar = c(3, 3, 0, 0), mgp = c(1, 0.3, 0), tck = -0.01),
  matplot_args_all = list(),
  matplot_args_group = list(),
  plot_profiles = TRUE,
  plot_stacked = TRUE,
  filter_info = list(gradient.max = c(0, 0.01)),
  par_val_ref = NULL,
  col_sub = c(total = rgb(0, 0, 0))
)
}
\arguments{
\item{sim_summary}{Output object from summarize_MCBE}

\item{nm_par}{Name of parameter to plot profile}

\item{nm_obj}{Name of object where par should be found in sim_summary. By default,
plot_profile will look in parms and parm.cons and use the first instance of nm_par it finds.
It works with data frames in sim_summary like parms and is programmed to work
with parm.cons as a special case, but may not work with other objects within sim_summary.}

\item{like_pattern}{pattern to use to identify likelihood components from \code{like} object in sim_summary.
Internally, the function uses \code{grepl} to identify names in \code{like}. The default "^lk."
uses regular expression to identify all likelihood components.}

\item{like_groups}{names of groups to use when plotting grouped likelihoods. The function adds "^lk."
to the group names to find parameters to group in \code{like}.}

\item{like_shift}{method to use to scale likelihood components for plotting:
"min" shifts each likelihood component by the overall minimum,
"min_par" shifts each likelihood component by its own minimum,
"none" returns raw likelihoods}

\item{like_exclude}{Names of columns in \code{like} to exclude from plotting profiles}

\item{like_exclude_stacked}{Names of columns in \code{like} to exclude from plotting stacked barplot}

\item{like_plot_last}{Name of the likelihood component to
plot last so that it is on top in the plots.}

\item{par_args}{arguments to pass to \code{par}}

\item{matplot_args_all}{arguments to pass to \code{matplot} for plot of all likelihoods}

\item{matplot_args_group}{arguments to pass to \code{matplot} for plot of grouped likelihoods}

\item{plot_profiles}{logical. Plot likelihood profiles as lines, possibly shifted with like_shift.}

\item{plot_stacked}{logical. Plot likelihood profiles as stacked areas (\code{ggplot})
columns not in like_prop_exclude.}

\item{filter_info}{list of ranges used to filter out sim runs. set to NULL to retain all results.}

\item{par_val_ref}{reference value to add to profile plots (e.g. parameter value from base model)}

\item{col_sub}{Specify colors to substitute for colors in the ungrouped profile plot.
named vector where the names match the names match the names in the plot legend.}
}
\description{
Plot results from MCBE parameter profile analysis
}
\examples{
\dontrun{
}
}
\keyword{MCBE}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
