% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_demo.R
\name{pop_demo}
\alias{pop_demo}
\title{pop_demo}
\usage{
pop_demo(
  Fi,
  age,
  h = NULL,
  R0 = NULL,
  rec_sigma = 0,
  M,
  sel_L,
  sel_D = NULL,
  PS,
  w,
  ep,
  P_st = 0.5,
  plus_group = TRUE,
  N_ast_ob = NULL,
  R_neq = NA,
  plots = FALSE,
  plot_digits = 3
)
}
\arguments{
\item{Fi}{Fishing mortality. numeric vector of length 1.}

\item{age}{age classes. numeric vector}

\item{h}{Beverton-Holt steepness parameter}

\item{R0}{Beverton-Holt R0 parameter. Numbers of fish at age-a (often age-0 or age-1).}

\item{rec_sigma}{recruitment standard deviation in log space. used to compute lognormal bias correction -- exp(sigma^2/2)}

\item{M}{Natural mortality rate}

\item{sel_L}{Selectivity at age for landings}

\item{sel_D}{Selectivity at age for discards. If NULL, assumed to be zero for all ages.}

\item{PS}{Proportion of fish to include in stock calculation, at age (e.g. mature females at age)}

\item{w}{weight at age in any units, used to convert numbers-at-age to weight-at-age. Results in weight will be in these units.}

\item{ep}{egg production proxy at age (e.g. weight, fecundity)}

\item{P_st}{Time of year when spawning occurs, as a proportion.}

\item{plus_group}{Passed to bamExtras::exp_decay (Should the function include a plus group? logical)}

\item{N_ast_ob}{Numbers of fish at age, at the time of spawning, observed. Allows the user to provide age structure of the population in the reference year (e.g. terminal year) used to compute observed stock size (i.e. numerator in stock status calculations). If values are provided, additional outputs will be computed in output data, with suffix "ob".}

\item{R_neq}{Recruitment used to scale non-equilibrium per recruit quantities to compute SPR-based reference points (e.g. stock size at Fi; SSB_40\%). An estimate of mean recruitment is often used here.}

\item{plots}{logical. Draw plots?}

\item{plot_digits}{number of significant digits to show in plot}
}
\description{
Computes population demographics at age and totals across age, for a single value of fishing mortality (F).
This function is called by \code{bamExtras::F_calc} to compute values over a range of F and compute reference points.
\code{pop_demo} is useful to return and plot the demographic data-at-age for specific values of F (e.g. Fmsy)
}
\examples{
# Example of how to use this function

}
\author{
Nikolai Klibansky, Erik Williams, and Kyle Shertzer
}
\keyword{assessment}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
