% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{random_walk}
\alias{random_walk}
\title{Simulate a random walk}
\usage{
random_walk(x = 1:100, y0 = 100, sd = 1, mu = 0)
}
\arguments{
\item{x}{independent variable values (e.g. vector of time steps)}

\item{y0}{Initial value of y}

\item{sd}{Standard deviation of normal error distribution}

\item{mu}{multiplicative drift parameter}
}
\description{
Simple function to simulate a random walk
}
\examples{
\dontrun{
plot(random_walk(),type="l")
plot(random_walk(),random_walk(),type="l")
# Random walks can be pretty
x <- 1:1000
y1 <- random_walk(x)
y2 <- random_walk(x)
plot(y1,y2,type="l")
points(y1,y2,type="p",pch=16,col=rainbow(length(x)))
}
}
