% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_AtlanticMenhaden}
\alias{rdat_AtlanticMenhaden}
\title{BAM stock assessment results (rdat file): Atlantic Menhaden (SEDAR 69)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 69 Stock Assessment Report Atlantic Menhaden
(\href{http://sedarweb.org/docs/sar/SEDAR\%2069\%20SAR\%20final\%20combined\%20benchmark.pdf}{PDF})
}
\usage{
rdat_AtlanticMenhaden
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
