% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_BlackSeaBass}
\alias{rdat_BlackSeaBass}
\title{BAM stock assessment results (rdat file): US Southeast Atlantic Black Sea Bass (SEDAR 56)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 76 Stock Assessment Report South Atlantic Black Sea Bass
(\href{https://sedarweb.org/documents/sedar-76-stock-assessment-report-south-atlantic-black-sea-bass/}{PDF})
}
\usage{
rdat_BlackSeaBass
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
