% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_BluelineTilefish}
\alias{rdat_BluelineTilefish}
\title{BAM stock assessment results (rdat file): Atlantic Blueline Tilefish (SEDAR 50)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 50 Stock Assessment Report Atlantic Blueline Tilefish
(\href{http://sedarweb.org/docs/sar/S50_AtlBLT_SAR_10.11.2017_FINALcorrected.pdf}{PDF})
}
\usage{
rdat_BluelineTilefish
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
