% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_Cobia}
\alias{rdat_Cobia}
\title{BAM stock assessment results (rdat file): Atlantic Cobia (SEDAR 58)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 58 Stock Assessment Report Atlantic Cobia
(\href{http://sedarweb.org/docs/sar/SEDAR\%2058\%20Final\%20SAR\%201.6.2020_revised\%209.14.2020.pdf}{PDF})
}
\usage{
rdat_Cobia
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
