% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_GagGrouper}
\alias{rdat_GagGrouper}
\title{BAM stock assessment results (rdat file): South Atlantic Gag (SEDAR 71)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 71 Stock Assessment Report South Atlantic Gag
(\href{http://sedarweb.org/docs/sar/SEDAR_71_SAR_4.19.21_final_withaddendum.pdf}{PDF})
}
\usage{
rdat_GagGrouper
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
