% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_GrayTriggerfish}
\alias{rdat_GrayTriggerfish}
\title{BAM stock assessment results (rdat file): South Atlantic Gray Triggerfish (SEDAR 82)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 82 Stock Assessment Report South Atlantic Gray Triggerfish
(\href{https://sedarweb.org/documents/sedar-82-south-atlantic-gray-triggerfish-final-stock-assessment-report/}{PDF})
}
\usage{
rdat_GrayTriggerfish
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
