% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_GreaterAmberjack}
\alias{rdat_GreaterAmberjack}
\title{BAM stock assessment results (rdat file): South Atlantic Greater Amberjack (SEDAR 59)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 59 Stock Assessment Report South Atlantic Greater Amberjack
(\href{http://sedarweb.org/docs/sar/SEDAR59_fullSAR_3.31.2020.pdf}{PDF})
}
\usage{
rdat_GreaterAmberjack
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
