% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_RedGrouper}
\alias{rdat_RedGrouper}
\title{BAM stock assessment results (rdat file): South Atlantic Red Grouper (SEDAR 53)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 53 Stock Assessment Report South Atlantic Red Grouper
(\href{http://sedarweb.org/docs/sar/S53_SA_RG_SAR_2.22.2017_0.pdf}{PDF})
}
\usage{
rdat_RedGrouper
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
