% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_RedPorgy}
\alias{rdat_RedPorgy}
\title{BAM stock assessment results (rdat file): South Atlantic Red Porgy (SEDAR 60)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 60 Stock Assessment Report South Atlantic Red Porgy
(\href{http://sedarweb.org/docs/sar/SEDAR_60_SA_Red_Porgy_SAR_4.9.2020.pdf}{PDF})
}
\usage{
rdat_RedPorgy
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
