% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_RedSnapper}
\alias{rdat_RedSnapper}
\title{BAM stock assessment results (rdat file): South Atlantic Red Snapper (SEDAR 73)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 73 Stock Assessment Report South Atlantic Red Snapper
(\href{http://sedarweb.org/docs/sar/SEDAR73_SARedSnapper_FullSAR_V3_0.pdf}{PDF})
}
\usage{
rdat_RedSnapper
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
