% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_ScampGrouper}
\alias{rdat_ScampGrouper}
\title{BAM stock assessment results (rdat file): Atlantic Scamp Grouper (SEDAR 68)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 68 Stock Assessment Report Atlantic Scamp Grouper
(\href{https://sedarweb.org/documents/sedar-68-atlantic-scamp-final-stock-assessment-report/}{PDF})
}
\usage{
rdat_ScampGrouper
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
