% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_SnowyGrouper}
\alias{rdat_SnowyGrouper}
\title{BAM stock assessment results (rdat file): South Atlantic Snowy Grouper (SEDAR 36 Update 2020)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 36 Update (2020) Stock Assessment Report South Atlantic Snowy Grouper
(\href{http://sedarweb.org/docs/suar/2020_SEDAR36U_SAR_November2020_revisedjanuary2021.pdf}{PDF})
}
\usage{
rdat_SnowyGrouper
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
