% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_SpanishMackerel}
\alias{rdat_SpanishMackerel}
\title{BAM stock assessment results (rdat file): South Atlantic Spanish Mackerel (SEDAR 28)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 28 Stock Assessment Report South Atlantic Spanish Mackerel
(\href{http://sedarweb.org/docs/sar/S28_SAR_SASpMack_FinalWithPStar_5.16.2013.pdf}{PDF})
}
\usage{
rdat_SpanishMackerel
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
