% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_Tilefish}
\alias{rdat_Tilefish}
\title{BAM stock assessment results (rdat file): South Atlantic Tilefish (SEDAR 66)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 66 Stock Assessment Report South Atlantic Tilefish
(\href{http://sedarweb.org/docs/sar/SEDAR\%2066\%20assessment\%20report\%20final_v5_0.pdf}{PDF})
}
\usage{
rdat_Tilefish
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
