% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdat_to_cpars.r
\name{rdat_to_cpars}
\alias{rdat_to_cpars}
\title{Build DLMtool cpars list from bam rdat object}
\usage{
rdat_to_cpars(
  rdat,
  nsim,
  nyears = NULL,
  proyears,
  Mat_age1_max = 0.49,
  herm = NULL
)
}
\arguments{
\item{rdat}{BAM output rdat (list) object read in with dget()}

\item{nsim}{Number of simulations in operating model}

\item{nyears}{Number of years of historical data}

\item{proyears}{Number of projection years}

\item{Mat_age1_max}{Limit maximum value of proportion mature of first age class (usually age-1).}

\item{herm}{Is the species hermaphroditic? If "gonochoristic", use female maturity. If "protogynous", use a function of male and female maturity.}
}
\description{
Build DLMtool cpars list from bam rdat object
}
\examples{
\dontrun{
# Build DLMtool Stock (Stock-class object)
Stock_RedPorgy <- rdat_to_Stock(rdat_RedPorgy)

# Build DLMtool operating model (OM-class object)
OM_RedPorgy <- new("OM", Stock_RedPorgy, Generic_Fleet, Precise_Unbiased, Perfect_Imp)
# Copy OM
OM_RedPorgy1 <- OM_RedPorgy2 <- OM_RedPorgy
# Build cpars list
cpars <- rdat_to_cpars(rdat_RedPorgy, nsim=slot(OM_RedPorgy,"nsim"), nyears=slot(OM_RedPorgy,"nyears"), proyears=slot(OM_RedPorgy,"proyears"))

slot(OM_RedPorgy2,"cpars") <- cpars

## Run and plot simple management strategy evaluations (MSE)
# Without cpars
mse_out1 <- runMSE(OM_RedPorgy1)
NOAA_plot(mse_out1)
# With cpars
mse_out2 <- runMSE(OM_RedPorgy2)
NOAA_plot(mse_out2)
}
}
\author{
Nikolai Klibansky
}
\keyword{DLMtool}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
