% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdat_to_Data.r
\name{rdat_to_Data}
\alias{rdat_to_Data}
\title{Build DLMtool Data object from bam rdat object}
\usage{
rdat_to_Data(
  rdat,
  Data = new("Data"),
  herm = NULL,
  nsim = 1,
  genus_species = NULL,
  Region = "Southeast US Atlantic",
  Fref_name = "Fmsy",
  Rec = "bam_recruits",
  CAA_abb = "all",
  CAL_abb = "all",
  Ind_abb = "all",
  CV_vbK = 0.001,
  CV_vbLinf = 0.001,
  CV_vbt0 = 0.001,
  CV_Cat = NULL,
  Units = "thousand pounds",
  Mat_age1_max = 0.49,
  length_sc = 0.1,
  wla_sc = NULL,
  catch_sc = 1
)
}
\arguments{
\item{rdat}{BAM output rdat (list) object read in with dget()}

\item{Data}{DLMtool Data object to start with}

\item{herm}{Is the species hermaphroditic? If "gonochoristic", use female maturity. If "protogynous", use a function of male and female maturity.}

\item{nsim}{see \code{\link[MSEtool]{OM-class}}}

\item{genus_species}{Genus and species names separated by a space (e.g. "Centropristis striata").}

\item{Region}{see \code{\link[MSEtool]{Data-class}}}

\item{Fref_name}{Name of F reference as named in rdat$parms list (e.g. Fmsy, F30)}

\item{Rec}{see \code{\link[MSEtool]{Data-class}}. Set to "bam_recruits" to set this value to the vector of recruits estimated by bam for years where rec devs were estimated. Set to NULL to leave it empty.}

\item{CAA_abb}{Abbreviation for identifying the observed age comp matrix for the catch-at-age (CAA) slot. Names of age compositions in the BAM rdat comp.mats list are expected to follow the naming convention "acomp.abb.ob". Examples from SEDAR 53 Red Grouper: "CVT", "HB", "cH". Set to "all" or "none" to use all or none of the age comps, respectively.}

\item{CAL_abb}{Abbreviation for identifying the observed index for the catch-at-length (CAL) slot. Analogous to CAA_abb. Names of length compositions in the BAM rdat comp.mats list are expected to follow the naming convention "lcomp.abb.ob". Set to "all" or "none" to use all or none of the length comps, respectively.}

\item{Ind_abb}{Abbreviation for identifying the observed index of abundance for the Ind slot. Names of indices in the BAM rdat t.series matrix are expected to follow the naming convention "U.abb.ob". Examples from SEDAR 53 Red Grouper: "CVT", "HB", "cH". If multiple (valid) abb values are provided, the corresponding indices will be averaged (geomean) and restandardized to a mean of 1. Abbreviations that don't match any indices will be ignored. Set to "all" or "none" to use all or none of the indices, respectively.}

\item{CV_vbK}{see \code{\link[MSEtool]{Data-class}}}

\item{CV_vbLinf}{see \code{\link[MSEtool]{Data-class}}}

\item{CV_vbt0}{see \code{\link[MSEtool]{Data-class}}}

\item{CV_Cat}{see \code{\link[MSEtool]{Data-class}}}

\item{Units}{Units of catch. see \code{\link[MSEtool]{Data-class}}}

\item{Mat_age1_max}{Limit maximum value of proportion mature of first age class (usually age-0 or age-1). Models sometimes fail when maturity of first age class is too high (e.g. >0.5)}

\item{length_sc}{Scalar (multiplier) to convert length units including wla parameter. MSEtool examples seem to use cm whereas BAM uses mm.}

\item{wla_sc}{Scalar (multiplier) to convert wla parameter to appropriate weight units. In null, the function will try to figure out if the weight unit of wla was g, kg, or mt based on the range of the exponent. The wla parameter will also be scaled by 1/length_sc^wlb.}

\item{catch_sc}{Scalar (multiplier) for catch. BAM catch is usually in thousand pounds (klb). The default 1 maintains that scale.}
}
\description{
Build DLMtool Data object from bam rdat object
}
\examples{
\dontrun{
# Write example here
}
}
\author{
Nikolai Klibansky
}
\keyword{DLMtool}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
