% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdat_to_Obs.r
\name{rdat_to_Obs}
\alias{rdat_to_Obs}
\title{Build MSEtool Obs (observation) object from bam rdat object}
\usage{
rdat_to_Obs(
  rdat,
  Obs = Perfect_Info,
  Ind_abb = "all",
  CAA_abb = "all",
  CAL_abb = "all",
  Cobs = c(0.05, 0.05),
  catch_sc = 1,
  CAA_ESS_sc = 0.3,
  CAL_ESS_sc = 0.3
)
}
\arguments{
\item{rdat}{BAM output rdat (list) object read in with dget()}

\item{Obs}{MSEtool Obs object to start with}

\item{Ind_abb}{Abbreviation for identifying the observed index of abundance for the Ind slot. Names of indices in the BAM rdat t.series matrix are expected to follow the naming convention "U.abb.ob". Examples from SEDAR 53 Red Grouper: "CVT", "HB", "cH". If multiple (valid) abb values are provided, the corresponding indices will be averaged (geomean) and restandardized to a mean of 1. Abbreviations that don't match any indices will be ignored. Set to "all" or "none" to use all or none of the indices, respectively.}

\item{CAA_abb}{Abbreviation for identifying the observed age comp matrix for the catch-at-age (CAA) slot. Names of age compositions in the BAM rdat comp.mats list are expected to follow the naming convention "acomp.abb.ob". Examples from SEDAR 53 Red Grouper: "CVT", "HB", "cH". Set to "all" or "none" to use all or none of the age comps, respectively.}

\item{CAL_abb}{Abbreviation for identifying the observed index for the catch-at-length (CAL) slot. Analogous to CAA_abb. Names of length compositions in the BAM rdat comp.mats list are expected to follow the naming convention "lcomp.abb.ob". Set to "all" or "none" to use all or none of the length comps, respectively.}

\item{Cobs}{Range of CVs for landings data. Default set to value used in BAM modelsee \code{\link[MSEtool]{Obs-class}}}

\item{catch_sc}{Scalar (multiplier) for catch. BAM catch is usually in thousand pounds (klb). The default 1 maintains that scale.}

\item{CAA_ESS_sc}{Scalar multiplied by CAA_nsamp (in numbers of fish) to compute effective sample size of age composition data (CAA_ESS) when empirical estimates of effective sample size are not available. Default of 0.3 is based on MSEtool generic Obs objects.}

\item{CAL_ESS_sc}{Scalar multiplied by CAL_nsamp (in numbers of fish) to compute effective sample size of length composition data (CAL_ESS) when empirical estimates of effective sample size are not available. Default of 0.3 is based on MSEtool generic Obs objects.}
}
\description{
Build MSEtool Obs (observation) object from bam rdat object
}
\examples{
\dontrun{
# Build MSEtool Stock, Fleet, and Obs (Stock-class, Fleet-class, and Obs-class objects)
Stock_RedPorgy <- rdat_to_Stock(rdat_RedPorgy)
Fleet_RedPorgy <- rdat_to_Fleet(rdat_RedPorgy)
Obs_RedPorgy <- rdat_to_Obs(rdat_RedPorgy)

# Build MSEtool operating model (OM-class object)
OM_RedPorgy <- new("OM", Stock_RedPorgy, Fleet_RedPorgy, Obs_RedPorgy, Perfect_Imp)
# Run and plot simple management strategy evaluation (MSE)
mse_out <- runMSE(OM_RedPorgy)
NOAA_plot(mse_out)
}
}
\author{
Nikolai Klibansky
}
\keyword{MSEtool}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
