% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_VermilionSnapper}
\alias{rdat_VermilionSnapper}
\title{BAM stock assessment results (rdat file): South Atlantic Vermilion Snapper (SEDAR 55)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 28 Stock Assessment Report South Atlantic Vermilion Snapper
(\href{http://sedarweb.org/docs/sar/S55_SA_VS_SAR_FINAL_4.16.2018.pdf}{PDF})
}
\usage{
rdat_VermilionSnapper
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
