% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGBP.r
\name{ROGBP}
\alias{ROGBP}
\title{color function like rainbow() but without yellow (or indigo)}
\usage{
ROGBP(n)
}
\arguments{
\item{n}{number of colors}
}
\description{
color function like rainbow() but without yellow (or indigo)
}
\examples{
\dontrun{
n <- 7
myalpha <- 0.25
mycol <- ROGBP(n) # Opaque ROGBP colors
mycol_tr2 <- color_trans(mycol,alpha=myalpha) # Transparent ROGBP colors modifing mycol, using alpha argument from color_trans

par(cex=2,pch=16)
plot(1:n,rep(1,n),col=mycol,ylim=c(0,5))
points(1:n,rep(3,n),col=mycol_tr2)
}

}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
