% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_proj.r
\name{run_proj}
\alias{run_proj}
\title{run_proj}
\usage{
run_proj(
  rdat = NULL,
  args_run_bam = NULL,
  args_bam2r = NULL,
  stochastic = FALSE,
  pstar = NULL,
  styr_proj = NULL,
  nyb_rcn = list(L = 3, U = 3, R = 5, comp = 3),
  nyp_cur = 3,
  nyp = 5,
  nm_yr_p = c("endyr", "endyr_dev_rec", "endyr_rec_dev", "endyr_rec_phase2",
    "endyr_rec_spr", "styr_regs", "endyr_proj"),
  F_cur = NULL,
  F_proj = NULL,
  L_cur = NULL,
  ages = NULL,
  N_styr_proj = NULL,
  S_styr_proj = NULL,
  sel_L = NULL,
  sel_D = NULL,
  sel_tot = NULL,
  sel_F_flt = NULL,
  wgt_mt = NULL,
  wgt_L_klb = NULL,
  wgt_D_klb = NULL,
  wgt_F_flt_klb = NULL,
  len_F_flt_mm = NULL,
  reprod = NULL,
  M = NULL,
  spawn_time = NULL,
  SR_par = NULL,
  SR_method = "BH",
  age_error = NULL,
  project_bam = FALSE,
  args_ia = list(U_nm = NULL),
  model_obs = list(cv_L = function(x = 0.2, n = nyp) {
     rep(x, n)
 }, cv_D =
    function(x = 0.2, n = nyp) {
     rep(x, n)
 }, cv_U = function(x, n = nyp) {
    
    rep(x, n)
 }, nfish = function(x, n = nyp) {
     rep(x, n)
 }, ntrip = function(x, n
    = nyp) {
     rep(x, n)
 }),
  plot = FALSE,
  nm_comp_sfx = c(""),
  t_series_na_vals = c(-9999, -99999, -999999),
  key_lenprob = c(all = "^len.cv")
)
}
\arguments{
\item{rdat}{BAM output rdat (list) object read in with dget().}

\item{args_run_bam}{list of arguments passed internally to run_bam if any are provided}

\item{args_bam2r}{list of arguments passed internally to bam2r if any are provided}

\item{stochastic}{logical. Is this a stochastic projection? Mostly runs the same code
either way, but stochastic projections add normal errors to recruitment. Stochastic
setting is intended to be run with rdat input from MCBE.}

\item{pstar}{Value to use for applying pstar scaling of Fmsy. If set to NULL, it is not used and pstar code will not be run.}

\item{styr_proj}{start year (i.e. first calendar year) of projection}

\item{nyb_rcn}{number of years in calculations that include recent landings
(L) or discards and their cvs, index cvs (U), recruitment (R), or numbers of
samples in age or length compositions (comp). Fleets without L or D within the
last nyb_rcn years of the base model will not be projected forward. named list}

\item{nyp_cur}{number of years to maintain current conditions before implementing management}

\item{nyp}{number of years of projection. Must be >nyp_cur}

\item{nm_yr_p}{names of \code{init} objects to project forward by \code{nyp}}

\item{F_cur}{current fully selected fishing mortality rate}

\item{F_proj}{fully selected fishing mortality rate}

\item{L_cur}{current level of landings. (2026-01-30 NPK: Not yet implemented in run_proj but will be used to set catch in projections instead of F).}

\item{ages}{ages. numeric vector}

\item{N_styr_proj}{abundance at age in styr_proj. numeric vector}

\item{S_styr_proj}{spawning stock size at age in styr_proj. Often biomass in mt, sometimes eggs in n. numeric vector}

\item{sel_L}{selectivity at age to compute landings. numeric vector}

\item{sel_D}{selectivity at age to compute dead discards. numeric vector}

\item{sel_tot}{selectivity at age to compute Z (includes landings and discards). numeric vector}

\item{sel_F_flt}{list of selectivity vectors (by age) or matrices (by year,age) for each fleet associated with landings or discards.}

\item{wgt_mt}{weight at age of population in metric tons (mt). numeric vector}

\item{wgt_L_klb}{weight at age of landings in thousand pounds (klb). numeric vector}

\item{wgt_D_klb}{weight at age of discards in thousand pounds (klb). numeric vector}

\item{wgt_F_flt_klb}{list of weight vectors (klb by age) or matrices (by year,age) for each fleet associated with landings or discards. numeric vector or matrix}

\item{len_F_flt_mm}{list of length vectors (mm by age) or matrices (by year,age) for each fleet associated with landings or discards. numeric vector or matrix}

\item{reprod}{reproductive contribution at age to SSB. numeric vector}

\item{M}{natural mortality rate, at age}

\item{spawn_time}{time of year for peak spawning}

\item{SR_par}{list of parameters associated with the stock-recruit (SR) relationship.
Currently only works with a Beverton-Holt SR relationship for which the parameters are:
h = steepness of spawner-recruit function, R0 = virgin recruitment of spawner-recruit
function, Phi0 = virgin spawners per recruit, biascorr = bias correction}

\item{SR_method}{Spawner-recruit function BH = Beverton-Holt, R0 = model estimated
mean recruitment (R0), GM = constant geometric mean of nyb_rcn$R recent years of
t.series$recruits. By default, run_proj will try to use BH, or R0, but if it can't
find all the parameters, it will resort to GM. Note GM method has not been reviewed
and should be used with caution.}

\item{age_error}{age error matrix to use for the projection years. By default, the function uses the same age_error matrix from the base model}

\item{project_bam}{logical. After the projection is run, should the function build a new set of projected bam files?}

\item{args_ia}{list of arguments to pass to interim adjustment function. See Details section below.}

\item{model_obs}{list of optional arguments for incorporating observation error into simulated sampling when building a dat file when project_bam=TRUE. See Details section below. NOT CURRENTLY USED 2025-09-17.}

\item{plot}{logical. Produce plots of extended base model including projected data}

\item{nm_comp_sfx}{Character vector. Possible suffix added to fleet abbreviation in comps
(e.g. c("a","b") in SEDAR 82 gray triggerfish). Used to match comps with catch when projecting
comp data when \code{project_bam=TRUE}}

\item{t_series_na_vals}{Numeric vector of values to consider as NA in t.series object.}

\item{key_lenprob}{Optional list for specifying cv values for length-at-age,
for each set of length compositions. (e.g. key_lenprob = c("D.rHD"="len.cv.L", "L.rGN"="len.cv.L")). Use keyword 'all', as in default to set all
values to a particular value in parm.cons.}
}
\description{
Project BAM forward in time to conduct deterministic or stochastic projections.
}
\details{
\code{args_ia}:
\itemize{
\item{U_nm}{The fleet abbreviation for an index of abundance that you want to use to adjust F (e.g. sTV, sCT). If set to \code{NULL}, no interim adjustment will be used.}
\item{type}{Method used for computing F adjustment. "Uprop" computes the mean U from recent years, divided by U from
reference year. The reference year is the terminal year of the last full stock assessment. The set of recent years is computed
as the current projection year plus yr_U_lag. By default, type = "Uprop"}
\item{yr_U_lag}{A vector of negative integers used to compute the set of years used for computing an interim adjustment
from the reference index (comma separated example values: -1, -1:-3, -2:4). By default, yr_U_lag = -(2:4).}
\item{yr_ia_by}{Frequency of interim adjustments, in year units. Used to determine the years between stock
assessments when interim adjustments will be computed and applied to F. The interim adjustment years are computed as
\code{yrlim_p <- endyr + c(1,nyp); yr_ia <- seq(yrlim_p[1], yrlim_p[2], by = yr_ia_by)}. By default, yr_ia_by = 5}
\item{include_yrlim}{Logical. Should interim analysis be conducted in the first and last years of the projection period?
By default, include_yrlim_p = FALSE}
}
\code{model_obs}:
Values supplied to model_obs should be provided as a vector of function class objects of length 1 or length equal to the number of fleets associated with the data type. For cv_U, ntrip, and nsamp, these function can use the argument x, where x is the geometric mean of values for a fleet for the nyb_rcn years of the assessment.
\itemize{
\item{cv_L}{Function for computing cvs of landings during projection period.}
\item{cv_D}{Function for computing cvs of discards during projection period.}
\item{cv_U}{Function for computing cvs of indices during projection period.}
\item{ntrip}{Function for computing the number of fish sampled for comps during projection period.}
\item{nsamp}{Function for computing the number of trips sampled for comps during projection period.}
}
}
\examples{
\dontrun{
# Run deterministic projection with defaults
proj_VeSn <- run_proj(rdat_VermilionSnapper,plot=TRUE)

# Run deterministic projection with defaults
# and project bam inputs, by adding the bam files with args_bam2r
proj_VeSn <- run_proj(rdat_VermilionSnapper,
project_bam=TRUE,
args_bam2r = list(
dat_obj=dat_VermilionSnapper,
tpl_obj=tpl_VermilionSnapper,
cxx_obj=cxx_VermilionSnapper
))
# The projections work for most of the assessments
proj_BlSb <- run_proj(rdat_BlackSeaBass,plot=TRUE)
proj_GaGr <- run_proj(rdat_GagGrouper,plot=TRUE)
proj_RdGr <- run_proj(rdat_RedGrouper,plot=TRUE)
proj_RdSn <- run_proj(rdat_RedSnapper,plot=TRUE)
proj_VeSn <- run_proj(rdat_VermilionSnapper,plot=TRUE)
# Some need additional arguments to run correctly
proj_GrTr <- run_proj(rdat_GrayTriggerfish,
nm_comp_sfx = c("a","b"),
plot=TRUE,
key_lenprob = c("D.rHDs"="len.cv.L", "L.rGNs"="len.cv.L")
)
}
}
\author{
Kyle Shertzer, Erik Williams, and Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
