% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.r
\name{sections}
\alias{sections}
\title{Identify sections of a vector of consecutive values whose differences does not exceed a buffer}
\usage{
sections(x, buffer = 0)
}
\arguments{
\item{x}{numeric vector}

\item{buffer}{value compared with difference between consecutive values. If the absolute difference exceeds the buffer, the values are considered different and not part of the same section.}
}
\description{
Identify sections of a vector that have the same consecutive value
}
\examples{
\dontrun{
x <- sample(c(1,0),100,replace=TRUE)
sections(x=x)
}

}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
