% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_eq.r
\name{sr_eq}
\alias{sr_eq}
\title{sr_eq}
\usage{
sr_eq(SR_switch = "beverton_holt", steep, R0, BiasCor, spr_F0, spr_F)
}
\arguments{
\item{SR_switch}{Set which stock-recruit relationship to use. 1. "beverton_holt" = Beverton-Holt, 2. "ricker" = Ricker, 3. "none" = none (average recruitment).}

\item{steep}{Beverton-Holt steepness parameter. numeric vector}

\item{R0}{Beverton-Holt R0 parameter (virgin recruitment). Numbers of fish at first age (often age-0 or age-1). numeric vector}

\item{BiasCor}{bias correction. numeric vector}

\item{spr_F0}{spawners per recruit at F=0}

\item{spr_F}{spawners per recruit at F}
}
\description{
Spawner-recruit equilibrium function (Beverton-Holt or Ricker). Used inside the get_ref_pts function.
This is code from the Beaufort Assessment Model converted from ADMB to R.
}
\examples{
\dontrun{
}
}
\author{
Kyle Shertzer, Erik Williams, and Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
