% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_rdat.r
\name{standardize_rdat}
\alias{standardize_rdat}
\title{Change non-standard names of rdat objects to standard naming conventions. Compute standard objects.}
\usage{
standardize_rdat(
  rdat,
  separator_key = c(. = "_"),
  separator_key_x = names(rdat),
  fleet_replace = TRUE,
  fleet_key = list(sCT = c("CVT"), sTV = c("CVID", "Mcvt"), sVD = c("VID"), sBT =
    c("Mbft"), sBL = c("sM"), sVL = c("vll"), sFT = c("FST"), sAN = c("nad"), sAM =
    c("mad"), sAS = c("sad"), sJA = c("jai"), sME = c("mareco"), cDV = c("cD"), cHL =
    c("cH", "cHl"), cLL = c("cL"), cOT = c("cO"), cPT = c("cp", "cP"), cTW = c("cT",
    "cTw", "cHTR"), cGN = c("comm"), cRN = c("cRn"), cRS = c("cRs"), cBN = c("cBn"), cBS
    = c("cBs"), rHB = c("HB", "hb", "rHb"), rHD = c("hbd", "HBD"), rGN = c("GR", "mrip",
    "rGe", "rA")),
  fleet_key_x = c("parms", "parm.cons", "t.series", "comp.mats", "sel.age", "sel.parms"),
  parms_key = c(R0 = "BH.R0", M.msst = "M.constant", M.MSST = "M.constant"),
  a_series_key = c(mat.male.endyr = "mat.male", mat.fem.endyr = "mat.female",
    prop.female.endyr = "prop.female"),
  t_series_key = c(total.L.wgt.klb = "total.L.klb"),
  na_values = "-99999"
)
}
\arguments{
\item{rdat}{rdat (list) object read in with dget()}

\item{separator_key}{named vector indicating what separator(s) should be used in the rdat. By default, any "_" will be replaced with "." in names of rdat elements}

\item{separator_key_x}{names of elements in the rdat list to apply the fleet_key to.}

\item{fleet_replace}{Should fleet_key be applied to fleet_key_x to replace fleet_abbreviations in the rdat?}

\item{fleet_key}{list where values are patterns (character vector) to find in rdat list elements specified with fleet_key_x and the names are the replacements. Patterns are searched with regex to find these values at the beginning or end of character strings, followed or preceded by ".", or in the middle of a string preceded and followed by ".".}

\item{fleet_key_x}{names of elements in the rdat list to apply the fleet_key to.}

\item{parms_key}{named vector where values are patterns (character strings) to find in names(rdat$parms) and the names in parms_key are replacements}

\item{a_series_key}{named vector where values are patterns (character strings) to find in names(a.series$parms) and the names in a_series_key are replacements}

\item{t_series_key}{named vector where values are patterns (character strings) to find in names(t.series$parms) and the names in t_series_key are replacements}

\item{na_values}{Other values to change to NA}
}
\description{
Change non-standard names of rdat objects to standard naming conventions. Compute standard objects.
}
\examples{
\dontrun{
names(rdat_RedPorgy$a.series)
rdat_RedPorgy_std <- standardize_rdat(rdat_RedPorgy)
names(rdat_RedPorgy_std$a.series)
}
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
