% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise.R
\name{stepwise}
\alias{stepwise}
\title{stepwise}
\usage{
stepwise(x, par = NULL)
}
\arguments{
\item{x}{numeric vector (e.g. )}

\item{par}{named vector of parameters}
}
\description{
very simple step function used for estimating selectivity (i.e. independent proportions for each age or length). The function
structure allows it to be passed to \code{fn_sel} in \code{\link[bamExtras]{sel_comp}}.
}
\note{
This is really a dummy function to use as \code{fn_sel} in \code{\link[bamExtras]{sel_comp}}
}
\examples{
\dontrun{
x <- 1:10
parms <- setNames(round(lgs(x,1,median(x)),2),paste0("a",x))
stepwise(x=x,par=parms)
}
}
\author{
Nikolai Klibansky
}
