% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_legend.r
\name{text_legend}
\alias{text_legend}
\title{Wrapper for legend function to quickly label a plot with text}
\usage{
text_legend(
  legend,
  x = "topleft",
  bty = "n",
  adj = c(0.1, 0),
  cex = 1.3,
  text.col = rgb(0, 0, 0.3, 0.4),
  ...
)
}
\arguments{
\item{legend}{see \code{\link[base]{legend}}}

\item{x}{see \code{\link[base]{legend}}}

\item{bty}{see \code{\link[base]{legend}}}

\item{adj}{see \code{\link[base]{legend}}}

\item{cex}{see \code{\link[base]{legend}}}

\item{text.col}{see \code{\link[base]{legend}}}

\item{...}{pass additional arguments to see \code{legend}}
}
\description{
Wrapper for legend function to quickly label a plot with text
}
\examples{
x <- 1000
plot(rnorm(x),rnorm(x),col=rainbow(x),pch=16)
text_legend("random rainbow dots")
plot(rnorm(x),rnorm(x),col=heat.colors(x),pch=16)
text_legend("BAM!!",x="center",text.col=1,cex=4)
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
