% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseries_plot.r
\name{tseries_plot}
\alias{tseries_plot}
\title{Complete a length or age composition data frame so that all have desired dimensions. This is a modified version of comp_complete. The two functions should probably be combined with options to choose between sub-functions.}
\usage{
tseries_plot(
  D,
  DErr,
  ErrStyle = "bands",
  ErrUnits = "",
  NErrUnits = 2,
  ylim = NULL,
  plotLabel = NULL,
  ...
)
}
\arguments{
\item{D}{data frame of mean observations by year, where columns are different groups (e.g. fleets)}

\item{DErr}{error values for each value in D. Must be same dimensions as D}

\item{ErrStyle}{Style of displaying error around the central tendency time series D. Can be"bars" or "bands"}

\item{ErrUnits}{Units of DErr. Used only for displaying text on plot (e.g. "SE")}

\item{NErrUnits}{Multiplied by DErr to determine how far above and below D to to plot the errors}

\item{ylim}{range of y on plot}

\item{plotLabel}{character string for labeling plot}

\item{...}{pass additional arguments to matplot}
}
\description{
Complete a length or age composition data frame so that all have desired dimensions. This is a modified version of comp_complete. The two functions should probably be combined with options to choose between sub-functions.
}
\examples{
\dontrun{
tseries_plot()
rdat <- rdat_BlackSeaBass
tser <- rdat$t.series
U <- tser[,grepl("^U.*ob$",names(tser))]
Ucv <- tser[,grepl("^cv.U.",names(tser))]
tseries_plot(U,Ucv,xlab="year",ylab="index of abundance",plotLabel=rdat$info$species)
}

}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
