% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerability.R
\name{vulnerability}
\alias{vulnerability}
\title{Estimate age and length at 5\% and full (100\%) vulnerability (total selectivity) and retention (landings selectivity)}
\usage{
vulnerability(
  Vdata,
  retdata = NULL,
  Linf,
  K,
  t0,
  length_sc = 1,
  V_scaleToMax = FALSE,
  ret_scaleToMax = FALSE
)
}
\arguments{
\item{Vdata}{Vulnerability-at-age vector (total selectivity of landings and discards) with names indicating ages}

\item{retdata}{Retention-at-age vector (combined selectivity for landings) with names indicating ages}

\item{Linf}{Von Bertalanffy growth parameter length infinity}

\item{K}{Von Bertalanffy growth parameter K}

\item{t0}{Von Bertalanffy growth parameter theoretical age at age zero}

\item{length_sc}{Scalar (multiplier) to convert length units. MSEtool examples seem to use cm whereas BAM uses mm.}

\item{V_scaleToMax}{Should vulnerability be rescaled to max vulnerability?}

\item{ret_scaleToMax}{Should retention be rescaled to max retention?}
}
\description{
Estimate age and length at 5\% and full (100\%) vulnerability (total selectivity) and retention (landings selectivity)
}
\examples{
\dontrun{

}
}
\author{
Nikolai Klibansky
}
\keyword{DLMtool}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
