% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamdat2mat.r
\name{bamdat2mat}
\alias{bamdat2mat}
\title{Write BAM dat file values to a set of matrices and data frames}
\usage{
bamdat2mat(
  CommonName = NULL,
  bam = NULL,
  writeFiles = "",
  writeFiles_dir = paste("bamdat2mat", writeFiles, sep = "_"),
  prompt_me = TRUE
)
}
\arguments{
\item{CommonName}{Common name of species modeled in BAM files. Only used when accessing dat, tpl, and cxx character vectors named as e.g. dat_CommonName}

\item{bam}{Object returned from \code{bam2r}.}

\item{writeFiles}{File extension of files to write output data to. Current options include "csv".}

\item{writeFiles_dir}{Name of directory to write files to.}
}
\description{
The script converts objects defined in the DATA SECTION of BAM into a set of
data objects that are more easily read by humans.
}
\examples{
\dontrun{
# Read in any of the current BAM models
dm_AtMe <- bamdat2mat("AtlanticMenhaden")
dm_BlSB <- bamdat2mat("BlackSeaBass")
dm_BlTi <- bamdat2mat("BluelineTilefish")
dm_Cobi <- bamdat2mat("Cobia")
dm_GagG <- bamdat2mat("GagGrouper")
dm_GrTr <- bamdat2mat("GrayTriggerfish")
dm_GrAm <- bamdat2mat("GreaterAmberjack")
dm_ReGr <- bamdat2mat("RedGrouper")
dm_RePo <- bamdat2mat("RedPorgy")
dm_ReSn <- bamdat2mat("RedSnapper")
dm_SnGr <- bamdat2mat("SnowyGrouper")
dm_Tile <- bamdat2mat("Tilefish")
dm_VeSn <- bamdat2mat("VermilionSnapper")

# Write data objects to csv
dm_AtMe <- bamdat2mat("AtlanticMenhaden",writeFiles="csv")
}
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
