% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_trans.r
\name{color_trans}
\alias{color_trans}
\title{Make opaque color transparent}
\usage{
color_trans(x, alpha = 0.25)
}
\arguments{
\item{x}{numeric vector}
}
\description{
This function allows you to modify opaque colors you are already using and make transparent versions. It's handy for adding confidence bands around mean values.
}
\examples{
\dontrun{
n <- 7
myalpha <- 0.25
mycol <- rainbow(n) # Opaque rainbow
mycol_tr <- rainbow(n,alpha=myalpha) # Transparent rainbow using alpha argument
mycol_tr2 <- color_trans(mycol,alpha=myalpha) # Transparent rainbow modifing mycol, using alpha argument from color_trans

par(cex=2,pch=16)
plot(1:n,rep(1,n),col=mycol,ylim=c(0,5))
points(1:n,rep(2,n),col=mycol_tr)
points(1:n,rep(3,n),col=mycol_tr2)
}

}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
