% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_complete.r
\name{comp_complete}
\alias{comp_complete}
\title{Complete each data frame in  a list of comp data frames so that all have the same dimensions}
\usage{
comp_complete(
  comp_list,
  comp_data_n,
  xBinByDataframe = FALSE,
  n_tag = c(".n", ".nfish"),
  n_colname_new = c("ntrip", "nfish"),
  rownames_num = NULL,
  colnames_num = NULL,
  output_type = "prop",
  minusGroup = FALSE,
  plusGroup = FALSE,
  valsToNA = -99999
)
}
\arguments{
\item{comp_list}{list of composition matrices, where column names are bin values (e.g. length or age), and rows are unique observations (e.g. years)}

\item{comp_data_n}{data frame with annual sample sizes associated with each comp matrix}

\item{xBinByDataframe}{If true, x values are determined separately for each data frame (e.g. length comps may be binned separately by fleet)}

\item{n_tag}{tag(s) added to end of columns in comp_data_n which indicate sample sizes (e.g. ".n", ".nfish")}

\item{n_colname_new}{column name values to associate with n_tag in output}

\item{rownames_num}{Names of data rows for each comp matrix (e.g. years) desired in output.}

\item{colnames_num}{Names of data columns for each comp matrix (e.g. age bins, length bins) desired in output.}

\item{output_type}{Type of comps to output: "input"= format used in spreadsheets when submitting data, including ntrip and and nfish columns, "prop" = matrix of proportions, "nfish" = matrix of numbers of fish}

\item{valsToNA}{Values that should be replaced with NA (e.g. -99999)}
}
\description{
Complete each data frame in  a list of comp data frames so that all have the same dimensions
}
\examples{
\dontrun{
# Combine age comp matrices for Red Porgy
cm <- rdat_RedPorgy$comp.mats
cma <- cm[grepl("^acomp.*ob$",names(cm))]

cman <- rdat_RedPorgy$t.series[,gsub("ob$","n",names(cma))]
cmanfish <- rdat_RedPorgy$t.series[,gsub("ob$","nfish",names(cma))]

# Output proportions at age
comp_complete(cma,output_type="prop")

# Output numbers of fish at age
names(cma) <- gsub(".ob$","",names(cma))
comp_complete(cma,cmanfish,output_type="nfish",n_tag=".nfish")

# Output age comps in data input format
comp_complete(cma,cbind(cman,cmanfish),output_type="input")
}

}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
