% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_decay.r
\name{exp_decay}
\alias{exp_decay}
\title{Exponential decay}
\usage{
exp_decay(age, Z, N1 = 1, plus_group = FALSE)
}
\arguments{
\item{age}{ages at which to compute numbers alive. Can be fractional ages (e.g. daily, monthly) but}

\item{Z}{instantaneous mortality rate}

\item{N1}{number of individuals at first age (i.e. \code{age[1]}) at the beginning of the year}

\item{plus_group}{Should the function include a plus group? logical}
}
\description{
Compute numbers of individuals alive at time in a population experiencing exponential decay, allowing age-varying Z. Can be used to compute numbers at age at the beginning of the year.
}
\examples{
\dontrun{
exp_decay(age=1:10,Z=0.2,N1=100)
exp_decay(age=1:10,Z=0.2,N1=100,plus_group = TRUE)
exp_decay(age=c(2,5,10),Z=0.2,N1=100,plus_group = TRUE)
}
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
