% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_calc.r
\name{F_calc}
\alias{F_calc}
\title{F_calc}
\usage{
F_calc(
  ac,
  h = NULL,
  R0 = NULL,
  sigma = 0,
  M,
  Fmax = 1,
  F_n = 101,
  sel,
  PS,
  W,
  ep,
  Px = 0.4,
  plots = FALSE,
  plus_group = TRUE,
  plot_digits = 3
)
}
\arguments{
\item{ac}{age classes. numeric vector}

\item{h}{Beverton-Holt steepness parameter}

\item{R0}{Beverton-Holt R0 parameter. Numbers of fish at age-a (often age-0 or age-1).}

\item{sigma}{lognormal bias correction -- exp(sigma^2/2)}

\item{M}{Natural mortality rate}

\item{Fmax}{Maximum fishing mortality rate}

\item{F_n}{Number of fishing mortality rates to try}

\item{sel}{Selectivity at age}

\item{PS}{Proportion of fish to include in stock calculation, at age (e.g. mature females at age)}

\item{W}{weight at age}

\item{ep}{egg production proxy at age (e.g. weight, fecundity)}

\item{Px}{reference proportion of unfished spawning potential ratio (e.g. Px=.40)}

\item{plot_digits}{number of significant digits to show in plot}
}
\description{
Calculations of stock dependent quantities over a range of fishing mortalities (F). This code calculates spawning potential ratio (SPR), and then maximum sustainable yield (MSY) using equilibrium methods incorporating a stock-recruit relationship. Many of the calculations, particularly the MSY calculations, are based on code
written by Kyle Shertzer in a very similar function.
}
\examples{
rdat <- rdat_VermilionSnapper
pr <- rdat$parms
as <- rdat$a.series
F_calc(ac=as$age, h=pr$BH.steep, R0=pr$R0, M=as$M, sel=rdat$sel.age$sel.v.wgted.tot,
       PS=as$mat.female, W=as$weight, ep=as$wgt.klb,
       plots=TRUE)

# When there is no SRR, the function will still compute SPR without computing MSY-based equibrium values
rdat <- rdat_RedSnapper
pr <- rdat$parms
as <- rdat$a.series
F_calc(ac=as$age, M=as$M, sel=rdat$sel.age$sel.v.wgted.tot,
       PS=as$mat.female, W=as$weight, ep=as$wgt.klb,
       plots=TRUE)
}
\author{
Kyle Shertzer and Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{dynamics}
\keyword{fisheries}
\keyword{population}
\keyword{stock}
