% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.r
\name{geomean}
\alias{geomean}
\title{Geometric mean: logarithm form}
\usage{
geomean(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Simple geometric mean function using only is.finite values of log(x).
}
\examples{
\dontrun{
x <- c(1,2,10)
geomean(x)
geomean2(x)
mean(x)
}

}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
