% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean2.r
\name{geomean2}
\alias{geomean2}
\title{Geometric mean: product form}
\usage{
geomean2(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Simple geometric mean function.This function will fail when x is a large value.
The logarithm form (geomean) works somewhat better but can't handle negative values.
}
\examples{
x <- c(1,2,10)
geomean(x)
geomean2(x)
mean(x)
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
