% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bam.R
\name{plot_bam}
\alias{plot_bam}
\title{Plot results from Beaufort Assessment Model}
\usage{
plot_bam(
  spp = NULL,
  graphics.type = "pdf",
  draft = FALSE,
  use.color = TRUE,
  years_plot = NULL,
  default_args = list(x = quote(spp), DataName = "spp", draft = draft, use.color =
    use.color, graphics.type = graphics.type),
  Bound.vec.plots.args = list(),
  BSR.time.plots.args = list(),
  CLD.total.plots.args = list(),
  Cohort.plots.args = list(),
  Comp.plots.args = list(),
  Comp.yearly.plots.args = list(),
  Eq.plots.args = list(),
  F.time.plots.args = list(),
  Growth.plots.args = list(),
  Index.plots.args = list(),
  Landings.plots.args = list(),
  NFZ.age.plots.args = list(),
  Parm.plots.args = list(),
  PerRec.plots.args = list(),
  Phase.plots.args = list(),
  Selectivity.plots.args = list(),
  StockRec.plots.args = list()
)
}
\arguments{
\item{spp}{bam output rdat object}

\item{graphics.type}{from FishGraph help: 'a vector of graphics file types to which graphics are saved. When NULL, no plots are saved.'}

\item{draft}{from FishGraph help: 'modifies plots for use in a report. When FALSE main titles are omitted.'}

\item{use.color}{from FishGraph help: 'plots are made in grayscale when FALSE'}

\item{years_plot}{vector of years to plot in time series. If NULL, all years are plotted}
}
\description{
A generic plotting function that calls plotting functions from FishGraph and bamExtras.
}
\examples{
\dontrun{
# Run MCBE, writing files to dir_bam_sim
plot_bam(rdat_BlackSeaBass)
}
}
\keyword{MCBE}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
