% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bam.R
\name{plot_bam}
\alias{plot_bam}
\title{Plot results from Beaufort Assessment Model}
\usage{
plot_bam(
  spp = NULL,
  graphics.type = "pdf",
  draft = FALSE,
  use.color = TRUE,
  years_plot = NULL,
  FGPlot_args = list(x = quote(spp), DataName = "spp", draft = draft, use.color =
    use.color, graphics.type = graphics.type),
  windows_args = list(),
  Bound.vec.plots.args = list(),
  BSR.time.plots.args = list(BSR.references = list("Bmsy", "SSBmsy", "Rmsy")),
  CLD.total.plots.args = list(CLD.w.references = list("msy.klb", "msy.klb", "Dmsy.klb"),
    CLD.n.references = list("msy.num", "msy.num", "Dmsy.num")),
  Cohort.plots.args = list(),
  Comp.plots.args = list(),
  Comp.yearly.plots.args = list(print.neff = TRUE, print.n = TRUE),
  Eq.plots.args = list(F.references = list("Fmsy", "F30"), user.Eq =
    list("L.eq.wholeklb", "L.eq.knum", "D.eq.knum")),
  F.time.plots.args = list(F.references = list("Fmsy", "F30"), F.additional =
    c("F.F30.ratio")),
  Growth.plots.args = list(plot.all = TRUE),
  Index.plots.args = list(),
  Landings.plots.args = list(),
  NFZ.age.plots.args = list(user.plots = "N.age.mdyr"),
  Parm.plots.args = list(),
  PerRec.plots.args = list(user.PR = list("SPR", "ypr.lb.whole"), F.references =
    list("Fmsy", "F30")),
  Phase.plots.args = list(Xaxis.F = FALSE),
  Selectivity.plots.args = list(compact = TRUE),
  StockRec.plots.args = list()
)
}
\arguments{
\item{spp}{bam output rdat object}

\item{graphics.type}{from FishGraph help: 'a vector of graphics file types to which graphics are saved. When NULL, no plots are saved.'}

\item{draft}{from FishGraph help: 'modifies plots for use in a report. When FALSE main titles are omitted.'}

\item{use.color}{from FishGraph help: 'plots are made in grayscale when FALSE'}

\item{years_plot}{vector of years to plot in time series. If NULL, all years are plotted}

\item{FGPlot_args}{set of arguments to pass to all FishGraph plotting functions}
}
\description{
A generic plotting function that calls plotting functions from FishGraph and bamExtras.
}
\examples{
\dontrun{
# Make plots
plot_bam(rdat_BlackSeaBass)
}
}
\keyword{MCBE}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
