% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MCBE.R
\name{plot_MCBE}
\alias{plot_MCBE}
\title{Plot results from MCBE uncertainty analysis}
\usage{
plot_MCBE(
  sim_summary,
  dir_figs = "figs",
  filter_info = list(gradient.max = c(0, 0.01), Fmsy = c(0, 5), Fend.Fmsy.mean = c(0, 5),
    R.sigma.par = c(0.2, 1), R0_prob = c(0.005, 0.995))
)
}
\arguments{
\item{sim_summary}{Output object from summarize_MCBE}

\item{dir_figs}{name of directory that will be created to store figures}

\item{filter_info}{list of ranges used to filter out MCBE results. set to NULL to retain all results.}
}
\description{
Plot results from MCBE uncertainty analysis
}
\examples{
\dontrun{
# Run MCBE, writing files to dir_bam_sim
run_MCBE("GrayTriggerfish", dir_bam_sim="sim_GrTr")
# Summarize results of MCBE and assign to object
MCBE_GrTr <- summarize_MCBE(dir_bam_sim="sim_GrTr")
# Plot MCBE results
}
}
\keyword{MCBE}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
