% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdat_GrayTriggerfish}
\alias{rdat_GrayTriggerfish}
\title{BAM stock assessment results (rdat file): South Atlantic Gray Triggerfish (SEDAR 41)}
\format{
An object of class \code{"list"}
}
\source{
SEDAR 41 Stock Assessment Report South Atlantic Gray Triggerfish
(\href{http://sedarweb.org/docs/sar/S41_SA_GTF_SAR_Final_4.19.2016.pdf}{PDF})
}
\usage{
rdat_GrayTriggerfish
}
\description{
Data stored in the standard rdat file output by the Beaufort Assessment Model (BAM)
}
\keyword{datasets}
