% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bam.R
\name{run_bam}
\alias{run_bam}
\title{Run bam model}
\usage{
run_bam(
  CommonName = NULL,
  fileName = "bam",
  dir_bam = NULL,
  bam = NULL,
  dat_file = NULL,
  tpl_file = NULL,
  cxx_file = NULL,
  dat_obj = NULL,
  tpl_obj = NULL,
  cxx_obj = NULL,
  admb_switch = "-nox",
  admb2r_obj = admb2r.cpp,
  cleanup = list(del = c("*.r0*", "*.p0*", "*.b0*", "*.log", "*.rpt", "*.obj", "*.htp",
    "*.eva", "*.bar", "*.tds", "*.o", "tmp_admb", "variance", "*.dep", "*.hes", "*.tmp"))
)
}
\arguments{
\item{CommonName}{Common name of species associated with dat, tpl, and cxx files}
}
\description{
Calls a shell script to run a bam model
}
\examples{
\dontrun{

}
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
