% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bam.R
\name{run_bam}
\alias{run_bam}
\title{Run bam model}
\usage{
run_bam(
  CommonName = NULL,
  fileName = "bam",
  dir_bam = NULL,
  bam = NULL,
  dat_file = NULL,
  tpl_file = NULL,
  cxx_file = NULL,
  dat_obj = NULL,
  tpl_obj = NULL,
  cxx_obj = NULL,
  standardize = TRUE,
  unlink_dir_bam = TRUE,
  admb_switch = "-nox",
  admb2r_obj = admb2r.cpp,
  cleanup = list(del = c("*.r0*", "*.p0*", "*.b0*", "*.log", "*.rpt", "*.obj", "*.htp",
    "*.eva", "*.bar", "*.tds", "*.o", "tmp_admb", "variance", "*.dep", "*.hes", "*.tmp"))
)
}
\arguments{
\item{CommonName}{Common name of species associated with dat, tpl, and cxx files}

\item{fileName}{Name given to BAM files, not including file extensions.}

\item{dir_bam}{Name of directory to write BAM files to, relative to the working directory.}

\item{bam}{Output of \code{bam2r}.}

\item{dat_file}{dat file path}

\item{tpl_file}{tpl file path}

\item{cxx_file}{cxx file path}

\item{dat_obj}{dat file read in as a character vector with readLines(con=dat_file)}

\item{tpl_obj}{tpl file read in as a character vector with readLines(con=tpl_file)}

\item{cxx_obj}{cxx file read in as a character vector with readLines(con=cxx_file)}

\item{standardize}{Should \code{\link[bamExtras]{standardize_bam}} be run by the function before running the BAM}

\item{unlink_dir_bam}{Should \code{dir_bam} be deleted after this function is run?}

\item{admb_switch}{Character string pasted to fileName to build \code{run_command} when running BAM with \code{shell(run_command)}
(i.e. \code{run_command <- paste(fileName, admb_switch)})}

\item{admb2r_obj}{Character string containing admb2r C++ code, which is written with \code{base::writeLines} to \code{dir_bam}}

\item{cleanup}{List object written to \code{cleanup.bat} file in \code{dir_bam}.}
}
\description{
Creates a folder in a directory \code{dir_bam} to write BAM files including model-specific files
(.dat, .tpl, and .cxx) a standard \code{admb2r.cpp} file stored in \code{bamExtras}, and a standard
but customizable \code{cleanup.bat} file. Model-specific files are supplied by the user in any one of several different ways
(see Arguments). It temporarily changes the working directory to \code{dir_bam}
then calls a shell script to run the BAM model, and changes the working directory back to the previous path.
The function invisibly returns the rdat object created by the cxx file, which can be assigned an object name.
}
\examples{
\dontrun{
Run a bam model and assign rdat output to object
rdat_AtMe <- run_bam("AtlanticMenhaden")
rdat_BlSB <- run_bam("BlackSeaBass")
rdat_BlTi <- run_bam("BluelineTilefish")
rdat_Cobi <- run_bam("Cobia")
rdat_GagG <- run_bam("GagGrouper")
rdat_GrTr <- run_bam("GrayTriggerfish")
rdat_GrAm <- run_bam("GreaterAmberjack")
rdat_ReGr <- run_bam("RedGrouper")
rdat_RePo <- run_bam("RedPorgy")
rdat_ReSn <- run_bam("RedSnapper")
rdat_ScGr <- run_bam("ScampGrouper")
rdat_SnGr <- run_bam("SnowyGrouper")
rdat_Tile <- run_bam("Tilefish")
rdat_VeSn <- run_bam("VermilionSnapper")

}
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
