% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_bam.r
\name{standardize_bam}
\alias{standardize_bam}
\title{Standardize object names in bam files (dat, tpl, and cxx)}
\usage{
standardize_bam(dat.file, tpl.file, cxx.file)
}
\arguments{
\item{dat.file}{dat file path}

\item{tpl.file}{tpl file path}

\item{cxx.file}{cxx file path}
}
\description{
This script reads in BAM dat, tpl, and cxx files, and converts them to R objects (character vectors of each line of the code). It then runs a set of gsub() functions to
replace object names with preferred naming conventions and returns dat, tpl, and cxx character vectors to use to rewrite the corresponding files with writeLines)
}
\examples{
standardize_bam()
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
