% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_rdat.r
\name{standardize_rdat}
\alias{standardize_rdat}
\title{Change non-standard names of rdat objects to standard naming conventions. Compute standard objects.}
\usage{
standardize_rdat(
  rdat,
  parms.key = c(R0 = "BH.R0"),
  a.series.key = c(mat.male.endyr = "mat.male", mat.fem.endyr = "mat.female",
    prop.female.endyr = "prop.female"),
  t.series.key = c(total.L.wgt.klb = "total.L.klb")
)
}
\arguments{
\item{rdat}{rdat (list) object read in with dget()}
}
\description{
Change non-standard names of rdat objects to standard naming conventions. Compute standard objects.
}
\examples{
\dontrun{
names(rdat_RedPorgy$a.series)
rdat_RedPorgy_std <- standardize_rdat(rdat_RedPorgy)
names(rdat_RedPorgy_std$a.series)
}
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
