% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_MCBE.R
\name{summarize_MCBE}
\alias{summarize_MCBE}
\title{Summarize results from MCBE uncertainty analysis}
\usage{
summarize_MCBE(
  dir_bam_sim = "sim",
  dir_bam_base = "base",
  obj_collect = c("info", "parms", "like", "spr.brps", "mse", "a.series", "t.series",
    "N.age", "Z.age", "sel.age"),
  coresUse = NULL
)
}
\arguments{
\item{dir_bam_sim}{Name of directory where run_MCBE results files have been saved, relative to the working directory.}

\item{obj_collect}{Names of objects in MCBE rdat files to collect and summarize}

\item{coresUse}{Number of cores to use when running parallel processes}
}
\description{
Summarize results from MCBE uncertainty analysis
}
\examples{
\dontrun{
# Run MCBE, writing files to dir_bam_sim
run_MCBE("GrayTriggerfish", dir_bam_sim="GrTr_sim", dir_bam_base="GrTr_base")
# Summarize results of MCBE and assign to object
MCBE_GrTr <- summarize_MCBE(dir_bam_sim="sim_GrTr")
}
}
\keyword{MCBE}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
