% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_MCBE.R
\name{summarize_MCBE}
\alias{summarize_MCBE}
\title{Summarize results from MCBE uncertainty analysis}
\usage{
summarize_MCBE(
  dir_bam_sim = "sim",
  dir_bam_base = "base",
  nm_model = "",
  obj_collect = c("info", "parms", "like", "spr.brps", "mse", "a.series", "t.series",
    "parm.cons", "N.age", "Z.age", "sel.age"),
  obj_labels = list(parm.cons = c("init", "lower", "upper", "phase", "prior_mean",
    "prior_var", "prior_pdf", "out")),
  coresUse = NULL
)
}
\arguments{
\item{dir_bam_sim}{Name of directory where run_MCBE results files are stored, relative to the working directory.}

\item{dir_bam_base}{Name of directory where base model results are stored, relative to the working directory.}

\item{nm_model}{character string indicating the model name as part of the MCBE file names.
For example, "bam" is the \code{nm_model} in "1-bam.rdat". This optional argument should be specified if there
are .rdat files in \code{dir_bam_sim} other than the MCBE output files. Otherwise, this function
will try to read all .rdat files.}

\item{obj_collect}{Names of objects in MCBE rdat files to collect and summarize}

\item{obj_labels}{list of names for each value of each parameter in parm.cons}
}
\value{
list of summarized outputs, most of which are named similar to typical BAM output (rdat) files.
This list can be passed to \code{plot_MCBE} to make a variety of useful plots
}
\description{
Summarize results from MCBE uncertainty analysis
}
\examples{
\dontrun{
# Run MCBE, writing files to dir_bam_sim
run_MCBE("GrayTriggerfish", dir_bam_sim="GrTr_sim", dir_bam_base="GrTr_base")
# Summarize results of MCBE and assign to object
MCBE_GrTr <- summarize_MCBE(dir_bam_sim="sim_GrTr")
}
}
\keyword{MCBE}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
