% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseries_plot.r
\name{tseries_plot}
\alias{tseries_plot}
\title{Time series plot}
\usage{
tseries_plot(
  x,
  err,
  errType = "sd",
  errStyle = "bands",
  errUnits = "SE",
  nErrUnits = 2,
  errLim = c(0, Inf),
  plotLabel = NULL,
  matplot_args = list(),
  polygon_args = list(),
  col_band_alpha = 0.3,
  pt_buffer = 0.1
)
}
\arguments{
\item{x}{data frame of mean observations by year, where columns are different groups (e.g. fleets)}

\item{err}{error values for each value in x. Must be same dimensions as x}

\item{errType}{type of error provided: "cv"= coefficient of variation or "sd"=standard deviation. If "cv" then x is multiplied by err to compute sd.}

\item{errStyle}{Style of displaying error around the central tendency time series x. Can be"bars" or "bands"}

\item{errUnits}{Units of err. Used only for displaying text on plot (e.g. "SE")}

\item{nErrUnits}{Multiplied by err to determine how far above and below x to to plot the errors}

\item{errLim}{Range for limiting the plotting of error bars or bands.}

\item{plotLabel}{character string for labeling plot}

\item{matplot_args}{optional list of additional arguments passed to \code{matplot}, which plots x.}

\item{polygon_args}{optional list of additional arguments passed to \code{polygon}, which plots error bands when errStyle="bands".}

\item{col_band_alpha}{alpha (transparency) value between 0 and 1 to adjust opacity of color bands}
}
\description{
This is a simple plotting fucntion that is useful when plotting time series typically found in stock assessments (e.g. catch, indices)
}
\examples{
\dontrun{
rdat <- rdat_BlackSeaBass
tser <- rdat$t.series
U <- tser[,grepl("^U.*ob$",names(tser))]
Ucv <- tser[,grepl("^cv.U.",names(tser))]
tseries_plot(U,Ucv,xlab="year",ylab="index of abundance",plotLabel=rdat$info$species)
}

}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
