% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_age.r
\name{vb_age}
\alias{vb_age}
\title{Von Bertalanffy (VB) length to age function}
\usage{
vb_age(L, Linf, K, t0)
}
\arguments{
\item{L}{lengths to convert to ages}

\item{Linf}{length infinity}

\item{K}{growth coefficient}

\item{t0}{age (time) at length zero}
}
\description{
Converts age to length with a Von Bertalanffy function
}
\examples{
vb_age()
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
