% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_len.r
\name{vb_len}
\alias{vb_len}
\title{Von Bertalanffy (VB) age to length function}
\usage{
vb_len(Linf, K, t0, a)
}
\arguments{
\item{Linf}{length infinity}

\item{K}{growth coefficient}

\item{t0}{age (time) at length zero}

\item{a}{ages to convert}
}
\description{
Converts age to length with a Von Bertalanffy function
}
\examples{
vb_len()
}
\author{
Nikolai Klibansky
}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
